% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-arrest-rate.R
\name{add_arrest_rate}
\alias{add_arrest_rate}
\title{Adds arrest rate to JACUSA2 result object.}
\usage{
add_arrest_rate(result)
}
\arguments{
\item{result}{created by \code{read_result())} or \code{read_results())}.}
}
\value{
result object with arrest rate added.
}
\description{
Adds arrest rate to a JACUSA2 result object.
The result object must be the output of the following JACUSA2 methods: 
rt-arrest or lrt-arrest.
}
