% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-sub.R
\name{add_base_sub}
\alias{add_base_sub}
\title{Adds base substitution column to JACUSA2 result object.}
\usage{
add_base_sub(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string specifying column prefix. "arrest", "through", or "NULL", default: NULL.}
}
\value{
result object with base substitutions added.
}
\description{
There must be only one reference base. A->G is okay, but AG->G is NOT allowed! 
Make sure to filter \code{result} with \code{filter_by_alleles() TODO} to comply with this restriction.
}
\examples{
data(rdd)
result <- add_base_sub(rdd)
str(result[['base_sub']])
}
