% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-sub-ratio.R
\name{add_base_sub_ratio}
\alias{add_base_sub_ratio}
\title{Add base substitution ratio}
\usage{
add_base_sub_ratio(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
result object with base subsitution ratio field "base_sub[_\code{base_type}]" added.
}
\description{
Adds ratio for base substitution defined by "ref_base" column and base counts 
given by \code{base_type}. It is required that there is only one reference base per site.
A new column "base_sub_\code{base_type}" will be added to the result object.
}
\details{
Depending on your data and JACUSA2 command line options the following values are possible for 
\code{base_type}:
\itemize{
\item bases, 
\item arrest, or 
\item through
}
Additionally, if read/base stratification has been used additional columns with the suffix "_marked" will be
added.
}
\examples{
TODO
}
\seealso{
TODO
}
