% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{add_bc}
\alias{add_bc}
\title{Add observed base calls.}
\usage{
add_bc(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
result object with observed base calls added.
}
\description{
Adds observed base calls calculated for base column specified by \code{base_type}.
}
