% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_bc_change}
\alias{add_bc_change}
\title{Add base call change for RDD comparisons of a JACUSA2 result file.}
\usage{
add_bc_change(jacusa2, ref = 0, bc = jacusa2$bc, aa = FALSE)
}
\arguments{
\item{jacusa2}{List object created by \code{read_jacusa()}.}

\item{ref}{TODO 0 => use ref_base > 0 => condition in bc}

\item{bc}{List of Vector of observed base calls.}

\item{aa}{Logical indicates if missing fields should added automatically.}
}
\value{
Returns a JACUSA object list with the additional "bc_change" field.
}
\description{
\code{add_bc_change()} adds base call changes for gDNA vs. cDNA comparisons.
Make sure \code{add_bc()} has been called to populate "bc1" and "bc2" fields or 
set option "aa" (i.e. auto_add) to TRUE.
All necessary but missing fields will be added automatically, possibly 
overriden existing fields of the JACUSA list object.
}
\examples{
## populate necessary fields manually
jacusa2 <- add_bc_matrix(rdd)
jacusa2 <- add_bc(jacusa2)
jacusa2 <- add_bc_change(jacusa2)
## plot distribution of base changes
barplot(table(jacusa2$bc_change))

## populate necessary fields automatically
jacusa2 <- add_bc_change(jacusa2, ref = 1, aa = TRUE)
## plot distribution of base changes
barplot(table(jacusa2$bc_change))

}
