% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_bc_change_ratio}
\alias{add_bc_change_ratio}
\title{Add base call change ratio (e.g.: editing frequency) to a JACUSA2 object list.}
\usage{
add_bc_change_ratio(jacusa2, ref, bc = jacusa2$bc,
  bc_matrix = jacusa2$bc_matrix, aa = FALSE)
}
\arguments{
\item{aa}{Logical indicates if missing fields should added automatically.}

\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{bc1}{Vector of observed base calls for condition 1.}

\item{bc2}{Vector of observed base calls for condition 2.}

\item{bc_matrix2}{Base call matrix of list of base call matrices for condition 2.}
}
\value{
Returns a JACUSA list object the additional fields: "bc_change_ratio", 
"bc_change_ratio1", ... "bc_change_ratio{n2}", where n2 is the number of replicates 
of condition 2.
}
\description{
\code{add_bc_change_ratio()} adds base call change ratio (e.g.: editing frequency) 
for each replicate and an average over all replictes for gDNA vs. cDNA comparisons
to the JACUSA list object. Condition 2 is be cDNA data. 
Depending on the number (=n2) of replicates in condition 2, n2 + 1 fields are added:
\itemize{
 \item bc_change_ratio1 - base call change ratio of condition 1 vs. replicate 1 
 of condition 2
 \item ... - ...
 \item bc_change_ratio{n2} - base call change ratio of condition 1 vs. replicate n2
 of condition 2
 \item bc_change_ratio - average over bc_change_ratio{1, ..., n2}
}
}
\examples{
## add fields manually
jacusa <- add_bc_matrix(rdd)
jacusa <- add_bc(jacusa)
jacusa <- add_bc_change(jacusa)
jacusa <- add_bc_change_ratio(jacusa)
## plot a boxplot of editing frequencies for each base change
boxplot(tapply(jacusa$bc_change_ratio, jacusa$bc_change, c))

#' ## add fields automatically
jacusa <- add_bc_change_ratio(rdd, aa = TRUE)
## plot a boxplot of editing frequencies for each base change
boxplot(tapply(jacusa$bc_change_ratio, jacusa$bc_change, c))

}
