% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_bc_matrix}
\alias{add_bc_matrix}
\title{Add base call count matrix to JACUSA list object.}
\usage{
add_bc_matrix(jacusa, bc_str1 = get_bc_str4condition(jacusa, 1),
  bc_str2 = get_bc_str4condition(jacusa, 2))
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{bc_str1}{Vector or list of string vectors that encode base calls for 
condition 1.}

\item{bc_str2}{Vector or list of string vectors that encode base calls for 
condition 2.}
}
\value{
Returns a JACUSA list object with "bc_matrix1" and "bc_matrix2" fields 
that contain the respective base call count matrix.
}
\description{
\code{add_bc_matrix()} adds base call count matrix for both 
conditions to JACUSA list object. JACUSA list object must have the following 
fields set: "bases1*" and "bases2*". Each "basesij" field, where i = condition 
and j respective replicate number, encodes base call counts (A,C,G,T) that are 
separated by ",": 10,2,0,0
}
