% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-obs.R
\name{add_bc_obs}
\alias{add_bc_obs}
\title{Add observed base calls.}
\usage{
add_bc_obs(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
result object with observed base calls added.
}
\description{
Adds observed base calls calculated from \code{base_type} column.
}
\examples{
data(rdd)
result <- add_bc_obs(result)
str(result[["bases"]])

# for rt-arrest
data(rdd)
# add observed base calls for arrest reads
result <- add_bc_obs(result, "arrest")
# add observed base calls for through reads
result <- add_bc_obs(result, "through")
# only print observed base calls for 
# arrest and through reads
cols <- paste0(c("arrest", "through"), "_bc_obs")
str(result[[cols]])
}
