% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-ratio.R
\name{add_bc_ratio}
\alias{add_bc_ratio}
\title{Adds base call ratio to a JACUSA2 result object.}
\usage{
add_bc_ratio(result, base_type = "bases")
}
\arguments{
\item{result}{created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string specifying column prefix. "arrest", "through", or "NULL", default: NULL.}
}
\value{
result object with base call ratio added.
}
\description{
Calculates and adds base call ratio to a JACUSA2 result object.
TODO
}
\examples{
data(rdd)
# use default base calls to define ratio
result <- add_bc_ratio(rdd)
TODOstr(subset(result, select = grep("bc_ratio", names(result))))
}
