% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{add_cov}
\alias{add_cov}
\title{Add coverage}
\usage{
add_cov(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{redfsad_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
result object with read coverage added.
}
\description{
Adds coverage specific for base counts stored in column \code{base_type}.
}
\examples{
TODO
}
