% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Add coverage fields to JACUSA2 list object.}
\usage{
add_coverage(jacusa2)
}
\arguments{
\item{jacusa2}{data frame created by \code{read_jacusa()}.}
}
\value{
Returns the initial data frame with coverage field added.

# TODO @examples
## add coverage  to data
#jacusa2 <- add_coverage(rdd)
## filter/keep sites that have a total coverage of 30
#jacusa <- filter_coverage(jacusa, "cov", 30)
}
\description{
\code{add_coverage()} calculates and adds read coverage to list of sites.
}
