% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data-desc.R
\name{add_data_desc}
\alias{add_data_desc}
\title{Adds data description to JACUSA2 result object.}
\usage{
add_data_desc(result, condition_desc = c(),
  format = "\%CONDITION\%_\%REPLICATE\%")
}
\arguments{
\item{result}{object created by \code{read_result()}.}

\item{condition_desc}{Vector mapping condition to more verbose description}

\item{format}{character vector must contain:}
}
\value{
JACUSA2 result object with data description added
}
\description{
This adds a more descriptive name to the data, extending condition, replicate information.
}
