% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-ref-ratio.R
\name{add_non_ref_ratio}
\alias{add_non_ref_ratio}
\title{Adds non reference base ratio to JACUSA2 result object.}
\usage{
add_non_ref_ratio(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string specifying column prefix. "arrest", "through", or "NULL", default: NULL.}
}
\value{
result object with non reference base substitution ratio added.
}
\description{
Given column \emph{ref_base}, calculates \eqn{non-ref. BC ratio = (total BCs - ref. BCs) / total BCs}
using \eqn{non-ref. BCs = ref. BCs - total BCs}.
}
\examples{
data(rdd)
result <- add_non_ref_ratio(rdd)
str(result$non_ref_ratio)
}
