% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-read.R
\name{add_read_matrix}
\alias{add_read_matrix}
\title{Add read start, through, and end counts vector for each condition of a 
JACUSA result file.}
\usage{
add_read_matrix(jacusa, read_str1 = get_read_str4condition(jacusa, 1),
  read_str2 = get_read_str4condition(jacusa, 2))
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa}.}

\item{read_str1}{Vector or list of vectors of string encoding read start, 
through, and end counts for condtition 1.}

\item{read_str2}{Vector or list of vectors of string encoding read start, 
through, and end counts for condtition 2.}
}
\value{
Returns a JACUSA list object with additional "read_matrix1" and 
"read_matrix2" fields.
}
\description{
\code{add_read_matrix()} adds read start, through, and end count vector for each 
condition of a JACUSA result file to the initial JACUSA list object.
}
