% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_ref_base2bc}
\alias{add_ref_base2bc}
\title{Add reference base to observed base call field}
\usage{
add_ref_base2bc(jacusa2, ref_field = "ref_base")
}
\arguments{
\item{jacusa2}{object created by \code{read_result()}}

\item{ref_field}{String or integer}
}
\value{
jacusa2 object enriched by ref. base 2 bc field
}
\description{
\code{add_ref_base2bc()} TODO description
}
