% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rt-arrest-stat.R
\name{add_rt_arrest_stat}
\alias{add_rt_arrest_stat}
\title{TODO}
\usage{
add_rt_arrest_stat(jacusa, read_matrix1 = jacusa$read_matrix1,
  read_matrix2 = jacusa$read_matrix2, aa = FALSE)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa}.}

\item{read_matrix1}{read info matrix for condition 1.}

\item{read_matrix2}{read info matrix for condition 2.}

\item{aa}{Logical indicates if missing fields should added automatically.}
}
\value{
Returns a JACUSA list object with the "stat" field set
}
\description{
\code{add_rt_arrest_stat}
}
\examples{
#'
# TODO

}
