% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{add_sub}
\alias{add_sub}
\title{Add base substitution.}
\usage{
add_sub(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
result object with base substitutions added.
}
\description{
Adds base substitution column to a JACUSA2 result object.
There must be only one reference base. A->G is okay, but AG->G is NOT allowed! 
Make sure to filter \code{result} with \code{filter_by_alleles() TODO} to comply with this restriction.
}
\examples{
data(rdd)
result <- add_sub(rdd)
str(result[["base_sub"]])
}
