% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-ratio.R
\name{add_sub_ratio}
\alias{add_sub_ratio}
\title{Add base substitution ratio}
\usage{
add_sub_ratio(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
result object with base subsitution ratio field \code{sub[_base_type]} added.
}
\description{
Adds ratio for base substitution defined by \code{ref} column and base counts 
given by \code{base_type}. It is required that there is only one reference base per site.
A new column \code{sub[_base_type]} will be added to the result object.
}
\details{
Depending on your data and JACUSA2 command line options the following values are possible for 
\code{base_type}:
\itemize{
\item bases, 
\item arrest, or 
\item through
}
TODO
Additionally, if read/base stratification has been used additional columns with the suffix "_marked" will be
added.
}
\examples{
data(rdd)
result <- add_sub_ratio(rdd)
str(result[["sub_ratio"]])
}
