% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-summary.R
\name{add_summary}
\alias{add_summary}
\title{Adds a summary of a column in a JACUSA2 result object. (EXPERIMENTAL)}
\usage{
add_summary(result, column, aggregate, summarise, suffix = "")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{column}{character string specifing the column to use.}

\item{aggregate}{character string defining how to aggregate: "site" or "condition".}

\item{summarise}{function to be used for \code{column} to generate summary.}

\item{suffix}{optional character string to add to the name of the new summary column.}
}
\value{
result object with a new summary column added.
}
\description{
For an existing \code{column} in a JACUSA2 result object, a \code{summarise} function, and 
an \code{aggregate} cirteria ("site" or "condition"), a new summary column will be created.
}
\details{
The name of the new column is a combination of the following arguments:
\describe{
  \item{\code{aggregate}]}{"site" or "condition",}
  \item{\code{column}]}{the column name, and}
  \item{(\code{suffix})}{an optional suffix string.}
}
Each argument will be separated by: "_".
}
