% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{All}
\alias{All}
\title{Merge tibbles with columns holding logical values with AND to a vector.}
\usage{
All(d)
}
\arguments{
\item{d}{tibble with logical values}
}
\value{
logical tibble with columns merged with AND.
}
\description{
Each column holding values of logicals will be merged row-wise with AND.
}
