% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{Allnum}
\alias{Allnum}
\title{Merge tibbles with columns holding the sum of numeric values of a vector.}
\usage{
Allnum(d)
}
\arguments{
\item{d}{tibble with numeric values}
}
\value{
numeric tibble with summed columns.
}
\description{
Each column holding values of numbers will be summed row-wise.
}
