% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{Any}
\alias{Any}
\title{Merge tibbles with columns holding logical values with OR to a vector.}
\usage{
Any(d)
}
\arguments{
\item{d}{tibble with logical values}
}
\value{
logical tibble with columns merged with OR.
}
\description{
Each column holding values of logicals will be merged row-wise with OR.
}
