% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APOBEC1}
\alias{APOBEC1}
\title{call2_APOBEC1_APOBEC1YTHmut_RC14_C2T_call2_result.out}
\format{
TODO:
\itemize{
  \item TODO
}
}
\usage{
data(APOBEC1)
}
\description{
TODO
}
\references{
TODO
}
\keyword{datasets}
