% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-robust.R
\name{apply_cond}
\alias{apply_cond}
\title{Apply FUN to all tibbles of conditions}
\usage{
apply_cond(X, FUN, cores = 1, ...)
}
\arguments{
\item{X}{TODO}

\item{FUN}{TODO}

\item{cores}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
