% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{arrest_col}
\alias{arrest_col}
\title{Column name for arrest base counts for \code{suffix}.}
\usage{
arrest_col(suffix = NULL)
}
\arguments{
\item{suffix}{TODO}
}
\value{
string the represents the column name for arrest for \code{suffix}.
}
\description{
Column name for arrest base counts for \code{suffix}. 
\code{suffix} will be used a suffix to create the column name:
"arrest[_\code{suffix}]".
}
