% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-arrest-rate.R
\name{arrest_rate}
\alias{arrest_rate}
\title{Calculates arrest rate.}
\usage{
arrest_rate(arrest, through)
}
\arguments{
\item{arrest}{numeric vector of read arrest events.}

\item{through}{numeric vector of read through events.}
}
\value{
numeric vector of arrest rate(s).
}
\description{
Calculates arrest rate from read arrest and read through counts.
}
