% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{arrest_rate_col}
\alias{arrest_rate_col}
\title{Column name for arrest rate for \code{suffix}.}
\usage{
arrest_rate_col(suffix = NULL)
}
\arguments{
\item{suffix}{string to specifiy base columns to retrieve arrest rate. Default: NULL}
}
\value{
string the represents the column name for arrest rate with  marked \code{suffix}.
}
\description{
Column name for arrest ratio for \code{suffix}. 
\code{suffix} will be used to create the column name:
\code{arrest_rate[_suffix]}.
}
