% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{base_call}
\alias{base_call}
\title{Observed base calls.}
\usage{
base_call(bases)
}
\arguments{
\item{bases}{tibble of base call counts}
}
\value{
vector of observed base calls
}
\description{
Observed base calls.
}
