% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{base_call_non_ref}
\alias{base_call_non_ref}
\title{Calculate the most abundant non reference base call.}
\usage{
base_call_non_ref(ref, bases)
}
\arguments{
\item{ref}{string vector reference bases.}

\item{bases}{nx4 matrix or data frame of base call counts or ratio.}
}
\value{
string vector of most abundant non reference base call for \code{bases}.
}
\description{
Calculates the most abundant non reference base call. 
\code{bases} is expected to be a nx4 matrix or data frame.
}
