% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-count.R
\name{base_count}
\alias{base_count}
\title{Retrieves the total number of observed unique bases.}
\usage{
base_count(bc, ref = NULL)
}
\arguments{
\item{bc}{vector of strings with observed base calls.}

\item{ref}{vector of strings with reference bases}
}
\value{
numeric vector of total number of bases.
}
\description{
Retrieves the total number of observed unique bases.
If \code{ref} is provided, reference base will be considered for counting, e.g.:
At some site only base 'A' was observed but the reference was 'T'.
\code{base_count} will be:
\itemize{
 \item 1, when \code{ref} is NOT provided, or
 \item 2, when \code{ref} is provided.
}
}
\examples{
bc <- c("A", "AG", "C")
str(base_count(bc))

# use default observed base calls and condsider reference base
ref <- c("A", "A", "T")
str(base_count(bc, ref))
}
