% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-sub-ratio.R
\name{base_sub_ratio}
\alias{base_sub_ratio}
\title{Calculates base substitution ratio (e.g.: editing frequency).}
\usage{
base_sub_ratio(ref_base, bases, other_bc = NULL)
}
\arguments{
\item{ref_base}{vector of reference bases.}

\item{bases}{matrix of observed base call counts.}

\item{other_bc}{vector strings that represents observed base calls. Default: NULL, will be calculated from \code{bases}.}
}
\value{
vector of base callsubstitution ratios.
}
\description{
Calculates base substitution ratio (e.g.: editing frequency) for \code{ref_base} and base counts stored
in nx4 matrix \code{bases}.
\code{ref_base} must have only one reference base. A->G is okay, but AG->G is NOT allowed!
TODO what if more than 2 alleles, what does other_bc do
}
\examples{
ref_base <- c("A", "A", "T")
# Only one non-ref. base
bases <- matrix(c( 1,  0, 9, 0,
                   5,  0, 5, 0,
                   0, 10, 0, 0),
                ncol = 4, byrow = TRUE)
# show base substiution ratios for observed base calls
base_sub_ratio(ref_base, bases)

# > one non-ref. base
bases <- matrix(c( 1,  2, 9, 1,
                   5,  3, 5, 1,
                   2, 10, 2, 1),
                ncol = 4, byrow = TRUE)
# we are only interesected in A->G, T->C
other_bc <- c("G", "G", "C")
# show base substiution ratios for A->G, T->C
base_sub_ratio(ref_base, bases, observed_bc)
}
