% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-sub-ratio.R
\name{base_sub_ratio_col}
\alias{base_sub_ratio_col}
\title{Column name for base substitution ratio for \code{base_type}.}
\usage{
base_sub_ratio_col(base_type)
}
\arguments{
\item{base_type}{string TODO}
}
\value{
string the represents the column name for base substition ratio for \code{base_type}.
}
\description{
Column name for base substitution ratio for \code{base_type}. 
\code{base_type} will be used a suffix to create the column name:
"base_sub[_\code{base_type}]".
}
