% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-obs.R
\name{bc_obs}
\alias{bc_obs}
\title{Calculate observed base calls for base counts.}
\usage{
bc_obs(bases)
}
\arguments{
\item{bases}{nx4 matrix or data frame of base call counts.}
}
\value{
string vector of obseved base calls for \code{bases}.
}
\description{
Calculates observed base calls for base counts. 
\code{bases} is expected to be a nx4 matrix or data frame.
}
\examples{
bases <- tibble::tribble(
 ~A, ~C, ~G, ~T,
  1,  0,  0,  0,
  0,  1,  0,  0,
  0,  0,  1,  0,
  0,  0,  0,  1,
  1,  1,  1,  1,
)
str(bc_obs(bases))
}
