% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-obs.R
\name{bc_obs_col}
\alias{bc_obs_col}
\title{Column name for observed base calls for \code{base_type}.}
\usage{
bc_obs_col(base_type = "bases")
}
\arguments{
\item{base_type}{string TODO}
}
\value{
string the represents the column name for observed base calls for \code{base_type}.
}
\description{
Column name for observed base calls for \code{base_type}. 
\code{base_type} will be used a suffix to create the column name:
"bc_obs[_\code{base_type}]". Using \code{base_type = "bases"} 
will result in the field "bc_obs".
}
