% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-ratio.R
\name{bc_ratio}
\alias{bc_ratio}
\title{Calculate base call ratios for base counts.}
\usage{
bc_ratio(bases)
}
\arguments{
\item{bases}{nx4 matrix or data frame of base call counts.}
}
\value{
data frame of base call rations \code{bases}.
}
\description{
Calculates base call ratios for base counts. 
\code{bases} is expected to be a nx4 matrix or data frame.
}
\examples{
bases <- matrix(c(5, 0, 5, 0,  1, 1, 1, 1), byrow = TRUE, ncol = 4)
ratio <- bc_ratio(bases)
}
