% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-ratio.R
\name{bc_ratio_col}
\alias{bc_ratio_col}
\title{Column name for base call ratios for \code{base_type}.}
\usage{
bc_ratio_col(base_type = "bases")
}
\arguments{
\item{base_type}{string specifiying base column. Default: bases.}
}
\value{
string that represents the column name for base call ratios for \code{base_type}.
}
\description{
Column name for base call ratios for \code{base_type}. 
\code{base_type} will be used to create the column name:
\code{bc_ratio[_base_type]}. Using \code{base_type = "bases"} 
will result in the field "bc_ratio".
}
