% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-ratio.R
\name{bc_ratio_cols}
\alias{bc_ratio_cols}
\title{Column name for base call ratios for \code{base_type}.}
\usage{
bc_ratio_cols(base_type = "bases")
}
\arguments{
\item{base_type}{string TODO}
}
\value{
string the represents the column name for base call ratios for \code{base_type}.
}
\description{
Column name for base call ratios for \code{base_type}. 
\code{base_type} will be used a suffix to create the column name:
"bc_ratio[_\code{base_type}]". Using \code{base_type = "bases"} 
will result in the field "bc_ratio".
}
