% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-bc.R
\name{bc_str2bc_matrix}
\alias{bc_str2bc_matrix}
\title{Convert string encoded base call counts to base call count matrix.}
\usage{
bc_str2bc_matrix(bc_str, collapse = FALSE)
}
\arguments{
\item{bc_str}{Vector of strings or list of string vectors where base counts (A,C,G,T) 
are separated by ",".}

\item{collapse}{Logical indicates if base counts should be aggregated.}
}
\value{
Returns a matrix or a list of matrices of base counts.
}
\description{
\code{bc_str2bc_matrix()} converts base counts encoded string vectors to 
count matrices. Typically, the result of \code{get_bc_str4condition(jacusa, i)}, 
where i = 1 or 2 and \code{read_jacusa()}.
}
\examples{
## Extract sequencing info for condition 1
bc_str1 <- get_bc_str4condition(rdd, 1)
## Convert string encoded base counts to count matrices
bc_matrix1 <- bc_str2bc_matrix(bc_str1)

}
