% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-max-alleles.R
\name{check_max_alleles}
\alias{check_max_alleles}
\title{Checks if site contains less or equal number of allowed alleles.}
\usage{
check_max_alleles(result, max_alleles = 2, use_ref_base = TRUE)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{max_alleles}{numeric specifices maximal allowed alleles per site.}

\item{use_ref_base}{boolean indicating, if "ref_base" should be used to count alleles per site (Default: TRUE).}
}
\value{
boolean indicating if "max_alleles" per site is obeyed.
}
\description{
Return TRUE|FALSE if the JACUSA2 result object has the maximum number 
of allowed alleles per site.
}
