% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-sub.R
\name{clean_read_sub}
\alias{clean_read_sub}
\title{Transform read substitution.}
\usage{
clean_read_sub(subs)
}
\arguments{
\item{subs}{string vector of read substitutions.}
}
\value{
string vector of base substitution.
}
\description{
Transforms read substitution. JACUSA2 can stratify reads based on 
base substitutions "-B A2G". This function will transform "A2G" to "A->G".
}
\examples{
subs <- c("*", "A2G", "*")
clean_read_sub(subs)
}
