% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-result.R
\name{copy_attr}
\alias{copy_attr}
\title{TODO}
\usage{
copy_attr(src, dest)
}
\arguments{
\item{src}{TODO}

\item{dest}{TODO}
}
\value{
\code{dest} with attributes copied from \code{src}
}
\description{
TODO
}
