% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_col}
\alias{cov_col}
\title{Column name for coverage for \code{base_type}.}
\usage{
cov_col(base_type = "bases")
}
\arguments{
\item{base_type}{string TODO}
}
\value{
string the represents the column name for coverage for \code{base_type}.
}
\description{
Column name for coverage for \code{base_type}. 
\code{base_type} will be used a suffix to create the column name:
"coverage[_\code{base_type}]". Using \code{base_type = "bases"} 
will result in the field "coverage".
}
