% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{create_data}
\alias{create_data}
\title{Create data matrix for NMF}
\usage{
create_data(
  result,
  scores = c("score", "deletion_score", "insertion_score"),
  ref_context = "A",
  context = 2,
  rename_score = "call2_score",
  meta_cond = TRUE
)
}
\arguments{
\item{result}{JACUSA2 result object.}

\item{scores}{Vector of strings. Columns in JACUSA2 result object to use as scores.
Default: "score", "insertion_score", "deletion_score"}

\item{ref_context}{reference context of a potential site. Default: "A".}

\item{context}{number of nucleotides to consider around a potential site. Default: 2.}

\item{rename_score}{Rename column "score". Default: "call2_score".}

\item{meta_cond}{logical - split features based on meta conditions}
}
\value{
Data frame
}
\description{
Create data matrix from JACUSA2 output object for non-negative matrix factorization.
}
\seealso{
GenomicRanges::seqlengths how to set sequence information on GRanges object to preserve correct ranges.
}
