% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{expand_tag}
\alias{expand_tag}
\title{Expand tagged reads}
\usage{
expand_tag(result)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}
}
\description{
This will expand tagged reads and create new column called "not_untagged_reads".
}
