% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extend}
\alias{extend}
\title{Extend site.}
\usage{
extend(gr, left = 0, right = 0)
}
\arguments{
\item{gr}{object created by \code{read_result()} or \code{read_results()}.}

\item{left}{integer number of nucleotides to shift.}

\item{right}{integer number of nucleotides to shift.}
}
\value{
Extended region
}
\description{
Extend site. Make sure `seqlengths` are set for `gr`.
Otherwise, check `?GenomeInfoDb::seqlengths`.
}
