% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel-ratio.R
\name{extract_info}
\alias{extract_info}
\title{Extract indel counts from the JACUSA2 info.}
\usage{
extract_info(s, f)
}
\arguments{
\item{s}{character vector of the JACUSA2 info.}

\item{f}{character of one of the following fields: e.g. "del11", "del12", "ins21", "ins22"...}
}
\value{
data.frame of the extracted info.
}
\description{
Extract indel counts from the JACUSA2 info.
}
