% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine-ref.R
\name{extract_ref}
\alias{extract_ref}
\title{Extract reference base from observed base calls.}
\usage{
extract_ref(cond, bc, ref_cond)
}
\arguments{
\item{cond}{vector of numeric.}

\item{bc}{vector of string representing base calls.}

\item{ref_cond}{numeric representing the condition to use to define reference base.}
}
\value{
vector of reference base(s) defined by \code{ref_cond} from \code{bc}.
}
\description{
Extracts reference base from observed base calls. 
Can be used within an \code{group_by_site()} statement.
}
\examples{
cond <- c(1, 1, 2, 2)
bc <- c("A", "A", "AG", "AG")
extract_ref(cond, bc, 1)
}
