% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine-ref-base.R
\name{extract_ref_base}
\alias{extract_ref_base}
\title{Extract reference base from observed base calls.}
\usage{
extract_ref_base(condition, bc, ref_condition)
}
\arguments{
\item{condition}{vector of numeric.}

\item{bc}{vector of string representing base calls.}

\item{ref_condition}{numeric representing the condition to use to define reference base.}
}
\value{
vector of reference base(s) defined by \code{ref_condition} from \code{bc}.
}
\description{
TODO
}
\examples{
TODO
}
