% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel-ratio.R
\name{extract_score}
\alias{extract_score}
\title{Extract indel scores from the JACUSA2 info.}
\usage{
extract_score(s, f)
}
\arguments{
\item{s}{character vector of the JACUSA2 info.}

\item{f}{character of one of the following fields: "insertion_score", "deletion_score".}
}
\value{
vector of the extracted scores
}
\description{
Extract indel scores from the JACUSA2 info.
}
