% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_unpack_info}
\alias{fast_unpack_info}
\title{Fast unpack info.}
\usage{
fast_unpack_info(strings, names)
}
\arguments{
\item{strings}{Vector of string to unpack}

\item{names}{Vector os strings of keys to unpacks}
}
\value{
List of unpacked values
}
\description{
Fast unpack info.
}
