% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{filter_all_artefacts}
\alias{filter_all_artefacts}
\title{Filters all sites with an artefact}
\usage{
filter_all_artefacts(filter)
}
\arguments{
\item{filter}{vector of strings that contains artefact filter information.}
}
\value{
vector of logical.
}
\description{
Removes sites that sites that have been marked by any feature/artefact filter.
}
\examples{
data(rdd)
# remove sites that are marked by artefact filter "D"
grouped <- group_by_site(rdd)
filtered <- filter_by(grouped, filter_all_artefacts(filter))
str(filtered)
}
