% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_allele_count}
\alias{filter_allele_count}
\title{Filter sites by the number of alleles per site.}
\usage{
filter_allele_count(jacusa, max_alleles = 2)
}
\arguments{
\item{jacusa}{List object create by \code{read_jacusa()}.}

\item{max_alleles}{Integer number of maximal alleles per site.}
}
\value{
Returs JACUSA list object with sites where max_alleles is obeyed.
}
\description{
Removes sites that have too many alleles.
}
