% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{filter_artefact}
\alias{filter_artefact}
\title{Filters sites by artefacts.}
\usage{
filter_artefact(filter, artefacts = NULL)
}
\arguments{
\item{filter}{vector of strings that contains artefact filter information.}

\item{artefacts}{vector of characters that correspond to feature/artefact filters to be filtered, default: NULL - Filter all.}
}
\value{
vector of logical.
}
\description{
Removes sites that sites that have been marked by feature/artefact filter.
}
\examples{
data(rdd)
sum(filter_artefact(rdd$filter, c("D")))
}
