% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{filter_by}
\alias{filter_by}
\title{TODO}
\usage{
filter_by(result, ...)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{...}{forwared to internal \code{dplyr::filter()} statement.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}
