% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-allele_count.R
\name{filter_by_allele_count}
\alias{filter_by_allele_count}
\title{Filters sites by the number of alleles per site.}
\usage{
filter_by_allele_count(result, max_alleles = 2, use_ref_base = TRUE)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{max_alleles}{numeric specifies maximal allowed alleles per site.}

\item{use_ref_base}{boolean indicating if "ref_base" should be used to count alleles (Default: TRUE).}
}
\value{
result object with sites where \code{max_alleles} is obeyed.
}
\description{
Removes sites that have too many alleles.
}
