% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_arrest_stat}
\alias{filter_by_arrest_stat}
\title{Filter JACUSA2 object by p-value from rt-arrest method.}
\usage{
filter_by_arrest_stat(jacusa2, max_pvalue)
}
\arguments{
\item{jacusa2}{List object created by \code{read_result()}.}

\item{max_pvalue}{Numeric value that represents the maximal p-value.}
}
\value{
Returns data frame with sites with a stat <= max_pvalue.
}
\description{
\code{filter_by_arrest_stat} removes sites that have a p-value more than some threshold that has been provided by user.
}
