% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_call_score}
\alias{filter_by_call_score}
\title{Filter JACUSA2 object sites by test-statistic from call method.}
\usage{
filter_by_call_score(jacusa2, min_score)
}
\arguments{
\item{jacusa2}{data frame created by \code{read_result()}.}

\item{min_score}{Numeric value that represents the minimal test-statistic.}
}
\value{
Returns data frame with sites with a test-statistic >= min_score.
}
\description{
\code{filter_by_call_score()} removes sites that have a test-statistic less than some threshold that has been provided by user.
}
