% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-filter-info.R
\name{filter_by_filter_info}
\alias{filter_by_filter_info}
\title{Filters sites by artefacts.}
\usage{
filter_by_filter_info(result, artefacts)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{artefacts}{vector of characters that correspond to artefact filters}
}
\value{
result object with sites filtered by filter info
}
\description{
Removes sites that contain artefacts.
}
