% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-score.R
\name{filter_by_max_score}
\alias{filter_by_max_score}
\title{Restains sites with maximal score.}
\usage{
filter_by_max_score(result, max_score, column = "score")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{max_score}{numeric value that specifies the maximal score.}

\item{column}{character string that specifies the column of the score.}
}
\value{
result object with sites where score <= max_score.
}
\description{
Retains sites that have score <= max_score.
}
\examples{
data(HIVRT)
str(filter_by_max_score(HIVRT, max_score = 0.05, column = "pvalue"))
}
