% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-robust-arrest-event.R
\name{filter_by_robust_arrest_events}
\alias{filter_by_robust_arrest_events}
\title{Retains sites that contain an arrest event in all replicates in at least one condition.}
\usage{
filter_by_robust_arrest_events(result, suffix = NULL, ...)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{suffix}{string to specifiy base columns to filter on. Default: NULL}

\item{...}{passed to internal \code{group_by_site()}}
}
\value{
result object with sites where at least one condition contains the arrest event in all replicates.
}
\description{
Enforces that at least one condition contains an arrest event in all replicates. 
The result object must be the output of the following JACUSA2 methods: 
\emph{rt-arrest} or \emph{lrt-arrest}.
}
\examples{
data(HIVRT)
result <- add_arrest_rate(HIVRT)
str(filter_by_robust_arrest_events(result))
# filtereing requiring a minimal difference between mean arrest rates from both conditions
# mean(arrest_rate1) - mean(arrest_rate2) >= 0.1
#str(filter_by_robust_arrest_events(result), min_diff = 0.1)
}
