% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_call_stat}
\alias{filter_call_stat}
\title{Filter JACUSA list of sites by test-statistic from call method.}
\usage{
filter_call_stat(jacusa, stat)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{stat}{Numeric value that represents the minimal test-statistic.}
}
\value{
Returns List with sites with a test-statistic >= stat.
}
\description{
\code{filter_call_stat()} removes sites that have a test-statistic less than some threshold that has been provided by user.
}
