% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Filter sites by read coverage (total, per replicates etc.).}
\usage{
filter_coverage(jacusa, field, cov)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{field}{Vector of strings determines what coverage fileds should be 
carried used for filtering.}

\item{cov}{Vector or numeric value of the minimal read coverage. Must have 
same dimension as fields}
}
\value{
Returns JACUSA list object of sites filtered by minimal read coverage
 according to field and cov.
}
\description{
\code{filter_coverage()} filters sites by customizable read coverage restrictions. 
Use \code{add_coverage()} first to add coverage fields to list.
}
\details{
\itemize{
  \item cov -> total coverage
  \item cov1 or cov2 -> total coverage of condition 1 or 2
  \item covs1 or covs2 -> coverage of replicates from condition 1 or 2
}
}
\examples{
## Keep sites that have a total read depth of at least 10 reads in 
## condition 1 and each replicates of condition 2 has a minimal read coverage 
## of 5 reads. 
jacusa <- filter_coverage(rdd, field = c("cov1", "covs2"), cov = c(10, 5))

}
