% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_min_variant_count}
\alias{filter_min_variant_count}
\title{Filter JACUSA results of RDD calls and ensure that sits is covered by minimal 
number of variant reads.}
\usage{
filter_min_variant_count(jacusa, min_count = 2)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()} and \code{add_base()}.}

\item{min_count}{Numeric value that specifies the minimal number of variant reads in sample 2 or cDNA.}
}
\value{
Returns List of sites that have at least min_count variant reads in cDNA.
}
\description{
\code{filter_min_variant_count()} Filters JACUSA result files of gDNA vs. cDNA 
comparisons and enforces a minimal number of variant bases in cDNA.
Per default gDNA is expected to be sample 1 and sample 2 to be cDNA!
Make sure the list contain fields: "bc1" and "bc2". Those fields 
corresponds to the bases that have been called for the sample.
The fields can be populated by calling \code{add_bc()}.
}
\examples{
## Filters sites that have less than 2 variant reads in the cDNA sample. 
jacusa <- add_bc_matrix(rdd)
jacusa <- add_bc(jacusa)
jacusa <- filter_min_variant_count(jacusa, min_count = 2)

}
