% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_rec}
\alias{filter_rec}
\title{Filter JACUSA List object recursively.}
\usage{
filter_rec(jacusa, index)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{index}{Vector of sites that should be retained.}
}
\value{
Returns List with sites that are contained in vector f.
}
\description{
\code{filter_rec()} filters list of sites recursively.
}
\examples{
## calculate variant count
jacusa <- add_bc_matrix(rdd)
jacusa <- add_bc(jacusa)
variant_count <- get_variant_count(jacusa$bc1, jacusa$bc2, jacusa$bc_matrix2)
## create index of sites that contain at least 10 variant bases
m <- do.call(cbind, variant_count)
index <- apply(m, 1, function(x) all(x >= 10))
## filter data according to index
filtered <- filter_rec(jacusa, index)

}
