% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_robust_variants}
\alias{filter_robust_variants}
\title{Retains sites that contain the variant base in all replicates of at least one sample.}
\usage{
filter_robust_variants(jacusa)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}
}
\value{
Returns List with sites where at least one sample contains the variant base in all replicates.
}
\description{
\code{filter_robust_variants} Enforces that at least one sample contains the variant base in all replicates.
}
\examples{
jacusa <- add_bc_matrix(rdd)
jacusa <- filter_robust_variants(jacusa)

}
