% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_rt_arrest_stat}
\alias{filter_rt_arrest_stat}
\title{Filter JACUSA list of sites by p-value from rt-arrest method.}
\usage{
filter_rt_arrest_stat(jacusa, pvalue)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{pvalue}{Numeric value that represents the maximal p-value.}
}
\value{
Returns List with sites with a stat <= pvalue.
}
\description{
\code{filter_rt_arrest_stat} removes sites that have a p-value more than some threshold that has been provided by user.
}
