% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gather_repl}
\alias{gather_repl}
\title{Extract values from a structured column.}
\usage{
gather_repl(id, x, meta_cond = NULL)
}
\arguments{
\item{x}{structured column}

\item{meta_cond}{vector of meta conditions}

\item{ids}{vector}
}
\value{
extracted column
}
\description{
Extract values from a structured column.
}
