% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_allele_count}
\alias{get_allele_count}
\title{Returns the number of alleles from a vector of base calls.}
\usage{
get_allele_count(bc)
}
\arguments{
\item{bc}{vector of base calls.}
}
\value{
number of unique alleles.
}
\description{
Returns the number of alleles from a vector of bases.
For a vector such as c("AG", "A") it returns 2.
}
