% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_alleles}
\alias{get_alleles}
\title{Get number of alleles within a vector of bases}
\usage{
get_alleles(bc)
}
\arguments{
\item{bc}{vector of bases}
}
\value{
thu unique numner of observed alleles
}
\description{
Gives the number of alleles within a vectors of bases.
For a vector such as c("AG", "A") it returns 2.
}
