% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{get_arrest_rate}
\alias{get_arrest_rate}
\title{Retrieve arrest rate.}
\usage{
get_arrest_rate(result, suffix = NULL)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{suffix}{TODO}
}
\value{
numeric vector with arrest rates for base counts with \code{suffix}.
}
\description{
TODO
}
\examples{
TODO
}
