% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-sub.R
\name{get_base_sub}
\alias{get_base_sub}
\title{Retrive base substition from a result object.}
\usage{
get_base_sub(result, base_type = "bases")
}
\arguments{
\item{result}{created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string specifying column prefix. "arrest", "through", or "NULL", default: NULL.}
}
\value{
string vector of base substitutions for \code{base_type}.
}
\description{
Extracts base substitution for some \code{base_type} from a JACUSA2 result object..
}
