% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-sub-ratio.R
\name{get_base_sub_ratio}
\alias{get_base_sub_ratio}
\title{Retrieve base substitution ratio.}
\usage{
get_base_sub_ratio(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining base counts that were used to define base substitution. Default: bases.}
}
\value{
numeric vector of base substitutios ratios for \code{base_type} in \code{result}.
}
\description{
Retrieves ratio for base substitution stored under column "base_sub_\code{base_type}".
Use \code{add_base_sub_ratio()} to add this column, if it does not exist.
}
\examples{
data(rdd)
result <- add_base_sub_ratio(rdd)
str(get_sub_ratio(result))
}
\seealso{
\code{add_base_sub_ratio()}
}
