% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_bc_change}
\alias{get_bc_change}
\title{Calculate base changes}
\usage{
get_bc_change(ref_base, observed_bc, sep = BC_CHANGE_SEP,
  no_change = BC_CHANGE_NO_CHANGE)
}
\arguments{
\item{ref_base}{vector of reference bases.}

\item{observed_bc}{vector of base calls.}

\item{sep}{character vector: "ref_base"sep"observed_bcs".}

\item{no_change}{character vector: how to format no changes.}
}
\value{
Vector of base call changes.
}
\description{
Calculates and formats base changes for provided reference bases and observed base calls.
}
