% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_bc_change_ratio}
\alias{get_bc_change_ratio}
\title{Calculates base call change ratio (e.g.: editing frequency).}
\usage{
get_bc_change_ratio(ref_base, bc_matrix)
}
\arguments{
\item{ref_base}{Vector of reference bases.}

\item{bc_matrix}{Observed base call matrix.}
}
\value{
Returns vector of base call change ratios.
}
\description{
Calculates base call change ratio (e.g.: editing frequency).
}
