% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-obs.R
\name{get_bc_obs}
\alias{get_bc_obs}
\title{Retrieve observed base calls for \code{base_type}.}
\usage{
get_bc_obs(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining base counts. Default: bases.}
}
\value{
string vector with observed base calls for \code{base_type}.
}
\description{
Retrieves observed base calls for base counts stored under "\code{base_type}".
}
\examples{
TODO
}
