% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-ratio.R
\name{get_bc_ratio}
\alias{get_bc_ratio}
\title{Retrive base call ratio from a result object.}
\usage{
get_bc_ratio(result, base_type = "bases")
}
\arguments{
\item{result}{created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string specifying base column. Default: bases.}
}
\value{
numeric tibble of base call ratios for \code{base_type}.
}
\description{
Extracts base call ratio for some \code{base_type} from a JACUSA2 result object.
}
\examples{
data(rdd)
# use default bases to define ratio
ratio <- get_bc_ratio(rdd)
str(ratio)

data(HIVRT)
# use arrest bases to define ratio
ratio <- get_bc_ratio(rdd, "arrest")
str(ratio)
}
