% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_bc_str4condition}
\alias{get_bc_str4condition}
\title{Get string encoded base call counts for a condition from a JACUSA result file}
\usage{
get_bc_str4condition(jacusa, condition)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{condition}{Integer value: 1 or 2.}
}
\value{
Returns a JACUSA list object of string encoded base calls counts.
}
\description{
\code{get_bc_str4condition()} returns string encoded base call counts columns for one condition (1 or 2) from a JACUSA result file.
}
\examples{
## Read JACUSA result file hek293_untreated.out
## Extract string encoded base calls for condition 1. 
bc_str1 <- get_bc_str4condition(rdd, 1)

}
