% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_comp_bc_matrix}
\alias{get_comp_bc_matrix}
\title{Get complemenarty base call count matrix.}
\usage{
get_comp_bc_matrix(bc_matrix)
}
\arguments{
\item{bc_matrix}{Matrix of base call counts.}
}
\value{
Returns complementary base call matrix.
}
\description{
Get complemenarty base call count matrix.
}
