% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_comp_bcs}
\alias{get_comp_bcs}
\title{Invert vector of base calls.}
\usage{
get_comp_bcs(bcs)
}
\arguments{
\item{bcs}{List of vectors of base calls.}
}
\value{
Returns a List of vector of complementary base calls.
}
\description{
Inverts vector of base calls.
}
