% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_false_discovery_rate}
\alias{get_false_discovery_rate}
\title{Calculate false discovery rate (FDR) based on observed base call changes and 
expected base call changes.}
\usage{
get_false_discovery_rate(bc_change_tbl, expected = format_bc_change("A", "G"))
}
\arguments{
\item{bc_change_tbl}{Vector object created by \code{table()}.}

\item{expected}{String that represents the expected base call change.}
}
\value{
Returns the calculate FDR.
}
\description{
Calculate false discovery rate (FDR) based on observed base call changes and 
expected base call changes.
}
