% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_motif}
\alias{get_motif}
\title{Get sequence}
\usage{
get_motif(gr, src)
}
\arguments{
\item{gr}{GRanges object.}

\item{src}{BSgenome or XStringSet object.}
}
\value{
Vector of strings.
}
\description{
Get the sequence in `gr` from `src` - respects strand information!
}
