% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-ref-ratio.R
\name{get_non_ref_ratio}
\alias{get_non_ref_ratio}
\title{Retrive non reference ratio from a result object.}
\usage{
get_non_ref_ratio(result, base_type = "bases")
}
\arguments{
\item{result}{created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string specifying column prefix. "arrest", "through", or "NULL", default: NULL.}
}
\value{
numeric vector of non reference substitution ratios for \code{base_type}.
}
\description{
Retrives non reference ratio based on \code{ref} and base calls specified by \code{base_type}.
}
\examples{
data(rdd)
result <- add_non_ref_ratio(rdd)
get_non_ref_ratio(result)
}
