% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_primary_sites}
\alias{get_primary_sites}
\title{Get number of primary sites}
\usage{
get_primary_sites(result)
}
\arguments{
\item{result}{object create by \code{read_result()}}
}
\value{
TODO
}
\description{
TODO
}
