% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-read.R
\name{get_read_str4condition}
\alias{get_read_str4condition}
\title{Get string encoded read start, through, and end counts for a condition from a JACUSA 2.x result file.}
\usage{
get_read_str4condition(jacusa, condition)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa}.}

\item{condition}{Integer value: 1 or 2.}
}
\value{
Returns a list of strings that encode read start, through, and end counts.
}
\description{
\code{get_read_str4condition()} returns read through and reads end count columns for a sample (1 or 2) from a JACUSA 2.x result file.
}
