% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_ref_base2bc}
\alias{get_ref_base2bc}
\title{Calculates reference base to base call (base change).}
\usage{
get_ref_base2bc(ref_base, observed_bc)
}
\arguments{
\item{ref_base}{vector of reference bases.}

\item{observed_bc}{vector of base calls.}
}
\value{
vector of formatted base call changes.
}
\description{
Calculates and formats base changes for \code{ref_base} and \code{observed_bc}.
}
