% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ref_base2bc_tbl.R
\name{get_ref_base2bc_tbl}
\alias{get_ref_base2bc_tbl}
\title{Get distirbution This is a TODO}
\usage{
get_ref_base2bc_tbl(result, ref_field)
}
\arguments{
\item{result}{TODO}

\item{ref_field}{TODO}
}
\value{
TODO
}
\description{
This is a TODO
}
