% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_robust_variant_index}
\alias{get_robust_variant_index}
\title{Get index of variant sites that have a variant base contained in all 
replicates of at least one condition.}
\usage{
get_robust_variant_index(bc_matrix1, bc_matrix2)
}
\arguments{
\item{bc_matrix1}{Base call matrix of list of base call matrices for condition 1.}

\item{bc_matrix2}{Base call matrix of list of base call matrices for condition 2.}
}
\value{
Returns the index of sites where the variant base is contained in all 
replicates of one condition.
}
\description{
Get index of variant sites that have a variant base contained in all 
replicates of at least one condition.
}
