% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_site_count}
\alias{get_site_count}
\title{Returns the number of primary sites.}
\usage{
get_site_count(result)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}
}
\value{
numeric vector or numeric of number of sites.
}
\description{
Retrieves the number of primary sites.
}
