% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{get_sub}
\alias{get_sub}
\title{Retrive base substition from a result object.}
\usage{
get_sub(result, base_type = "bases")
}
\arguments{
\item{result}{created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
string vector of base substitutions for \code{base_type}.
}
\description{
Extracts base substitution for some \code{base_type} from a JACUSA2 result object..
}
\examples{
data(rdd)
result <- add_sub(rdd)
str(get_sub(result))
}
