% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_variant_count}
\alias{get_variant_count}
\title{Calculates the number of variant reads in JACUSA results files}
\usage{
get_variant_count(bc1, bc2, bc_matrix2)
}
\arguments{
\item{bc1}{Vector of base calls in condition 1.}

\item{bc2}{Vector of base calls in condition 2.}

\item{bc_matrix2}{Base call matrix or list of base call matrices.}
}
\value{
Returns List of count that represent the number of variant reads in condition 2
}
\description{
Calculates the number of variant reads in JACUSA result files of gDNA vs. cDNA comparisons.
Per default gDNA is expected to be condition 1 and cDNA to be condition 2!
}
