% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by-site.R
\name{group_by_site}
\alias{group_by_site}
\title{Group JACUSA2 result object by coordinates}
\usage{
group_by_site(result, opt_vars = c(), ...)
}
\arguments{
\item{result}{object created by \code{read_result(s)()}.}

\item{opt_vars}{optional variables to group by (need to be present)}

\item{...}{forward options to dplyr::group_by}
}
\value{
grouped JACUSA2 result object
}
\description{
TODO
}
