% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HIVRT}
\alias{HIVRT}
\title{JACUSA2 output for rt-arrest on HIVRT condition in Zhou et. al 2018 data}
\format{a tibble with 19 elements:
\itemize{
	\item contig: Character string representing the contig of the variant
	\item start: Numeric position of variant (>=0)
	\item end: Numeric corresponds to "start + 1"
	\item name: Character string. Currently, name of used method (\emph{call-\{1,2\}}, \emph{pileup}, \emph{rt-arrest}, or \emph{lrt-arrest})
	\item score: Numeric value representing the test-statistc. Higher values indicate more divergent pileups
	\item strand: Character representing strand information; "+", "-", or "."(no strand information available)
	\item info: Character string separated with ";" provding additional data for this specific site. Empty field is equal to "*"
	\item filter_info: ","-separated character string showing feature filter information. Empty field is equal to "*"
	\item ref_base: Character "A", "C", "G", "T", or "N" representing the reference base for this site - inverted when strand is "-".
	\item base_type: Character string inndicating what type of base calls are shown. Here: "total", "arrest_bases", or "through_bases".
	\item condition: Numeric value: "1"(here: +GMC) or "2"(-GMC)
	\item replicate: Numeric value: "1" or "2" for all conditions
	\item bc_A: Numeric value representing base call count for A
	\item bc_C: Numeric value representing base call count for C
	\item bc_G: Numeric value representing base call count for G
	\item bc_T: Numeric value representing base call count for T
	\item bc: Observed base calls for this site. Combination of: "A", "C", "G", and "T"
   \item cov: Numeric value indicating the read coverage for this site
}}
\usage{
data(HIVRT)
}
\description{
Zhou2018 et al. 2018 map RNA modification of pseudouridine \eqn{\Psi} by chemically modifying
pseudouridines with carbodiimide (+CMC) and detecting arrest events that are induced by reverse
transcription stops in high-throughput sequencing under 3 different conditions (HIVRT, SIIIMn, and
SIIIMg). 
Result are for running JACUSA2 rt-arrest on processed pairwise +CMC and -CMC(control) data in HIVRT condition are presented.
}
\details{
Check Section "Reverse transcriptase arrest events" in https://github.com/dieterich-lab/JACUSA2/manual/manual.pdf
for details on pre-processing and mapping primary sequencing data.
}
\references{
Zhou, K. I.; Clark, W. C.; Pan, D. W.; Eckwahl, M. J.; Dai, Q. & Pan, T.
Pseudouridines have context-dependent mutation and stop rates in high-throughput sequencing 
RNA Biology, Informa UK Limited, 2018 , 15 , 892-900
}
\keyword{datasets}
