% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HIVRT_call2}
\alias{HIVRT_call2}
\title{JACUSA2 output for call-2 on HIVRT condition in Zhou et. al 2018 data}
\format{
a tibble with 18 elements:
\itemize{
   \item id: Character string representing a unique identifier - created from contig, start, [end], and strand.
	\item contig: Character string representing the contig of the variant
	\item start: Numeric position of variant (>=0)
	\item end: Numeric corresponds to "start + 1"
	\item name: Character string. Name of used method \emph{call-2}
	\item score: Numeric value representing the test-statistc. Higher values indicate more divergent pileups
	\item strand: Character representing strand information; "+", "-", or "."(no strand information available)
	\item bases: Numeric tibble with representing counts for A, C, G, and T base calls.
   \item cov: Numeric value indicating the read coverage for this site
   \item backtrack1: Character - indicator if backtracking was used for condition 1.
   \item backtrack2: Character - indicator if backtracking was used for condition 2.
   \item backtrackP: Character - indicator if backtracking was used for condition pooled condition.
   \item reset1: Character - indicator if default estimation was unstable with for condition 1.
   \item reset2: Character - indicator if default estimation was unstable with for condition 2.
   \item resetP: Character - indicator if default estimation was unstable with for pooled condition.
	\item info: Character string separated with ";" provding additional data for this specific site. Empty field is equal to "*"
	\item filter: ";"-separated character string showing feature filter information. Empty field is equal to "*"
	\item ref: Character "A", "C", "G", "T", or "N" representing the reference base for this site - inverted when strand is "-".
}
}
\usage{
data(HIVRT_call2)
}
\description{
Zhou2018 et al. 2018 map RNA modification of pseudouridine \eqn{\Psi} by chemically modifying
pseudouridines with carbodiimide (+CMC) and detecting arrest events that are induced by reverse
transcription stops in high-throughput sequencing under 3 different conditions (HIVRT, SIIIMn, and
SIIIMg). 
Result are for running JACUSA2 call-2 on processed pairwise +CMC and -CMC(control) data in HIVRT condition are presented.
}
\details{
Check Section "Reverse transcriptase arrest events" in 
https://github.com/dieterich-lab/JACUSA2/blob/master/manual/manual.pdf
for details on pre-processing and mapping primary sequencing data.
}
\references{
Zhou, K. I.; Clark, W. C.; Pan, D. W.; Eckwahl, M. J.; Dai, Q. & Pan, T.
Pseudouridines have context-dependent mutation and stop rates in high-throughput sequencing 
RNA Biology, Informa UK Limited, 2018 , 15 , 892-900
}
\keyword{datasets}
