% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{id}
\alias{id}
\title{Merged coordinate information}
\usage{
id(result)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}
}
\value{
merged coorindates information
}
\description{
Merge contig:start|start-end:strand from AJCUSA2 result object
}
