% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_result}
\alias{import_result}
\title{Import JACUSA2 result file}
\usage{
import_result(
  file,
  unpack = TRUE,
  tmpdir = tempdir(),
  showProgress = FALSE,
  ...
)
}
\arguments{
\item{file}{String that represents the file name of the JACUSA2 output or an URL.}

\item{unpack}{Boolean unpack info field.}

\item{tmpdir}{String @see `data.table::fread`.}

\item{showProgress}{Boolean @see `data.table::fread`.}

\item{...}{parameters forwarded to `data.table::fread`.}
}
\value{
RangedExperiment object representing JACUSA2 result
}
\description{
\code{import_result()} reads data that was generated by JACUSA2 and creates a JACUSA2 result object.
}
