% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{JACUSA2VCF}
\alias{JACUSA2VCF}
\title{Helper function - create VCF
Convert JACUSA output to VCF format}
\usage{
JACUSA2VCF(jacusa, ref, invert = FALSE)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa}.}

\item{ref}{Vector of characters.}

\item{invert}{Logical indicates if on "-" strand complementary base calls 
should be calculated.}
}
\value{
Data frame that represents JACUSA list object in VCF file format.

#@export
}
\description{
This converts JACUSA output to VCF file format
}
