% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{package}
\name{JacusaHelper}
\alias{JacusaHelper}
\alias{JacusaHelper-package}
\title{Jacusa2Helper: A package for post-processing JACUSA 1.x and 2.x result files.}
\description{
The Jacusa2Helper package provides the following categories of important functions:
input/output, get, add, and filter to post-process result files of JACUSA 2.x.
}
\details{
\itemize{
 \item[call{1,2}]  calling variants from one condition or condition comparisons.
 \item[rt-arrest] identificaion of RNA modifications by reverse transcription arrest.
}

One major new feature of JACUSA 2.x is the identificaion of RNA modifications by reverse
transcription arrest.

When calling RDDs where the RNA-Seq sample has been generated with stranded sequencing library, 
base change(s) can be directly inferred and if necessary base calls can be inverted.
By default, DNA need to be provided as condition 1 and cDNA as condition 2!
Warning: Some function do not support replicates or are exclusively applicable on RDD or RRD 
result files!
use jacusa <- read_jacusa("jacusa.out") to read JACUS output and
\itemize{
  \item bc_str1 <- get_bc_str4condition(jacusa, 1) to extract string encoded base call counts for condition 1, or
  \item read_str1 <- get_read_str4condition(jacusa, 1) to extract string encoded read start, through, and end counts.
}

The central object in Jacusa2Helper is the JACUSA list object that represents
JACUSA result file "result.out" and can be created via \code{l <- read_jacusa("result.out")}.
}
\section{input/output functions}{

The functions read_jacusa, and write_jacusa facilitate input and output operations on JACUSA 1.x and 2.x output files.

See:
\itemize{
  \item read_jacusa
  \item write_jacusa
}
}

\section{get functions}{

This functions calculate TODO
}

\section{add functions}{

TODO
This functions calculate and add additional such as read depth or base call changes. 
This includes functions that are encoded as vectors of strings to count vectors or matrices.

See:
\itemize{
  \item add_bc
  \item add_bc_change
  \item add_bc_change_ratio
  \item add_coverage
  \item add_read
}
}

\section{filter functions}{

This function set enables processing of JACUSA 1.x and 2.x output files such as filtering by read coverage 
or enforcing a minimal number of variant base calls per sample.

See:
\itemize{
 \item filter_allele_count
  \item filter_call_stat
  \item filter_coverage
  \item filter_rt_arrest_stat
  \item filter_robust_variants
  \item filter_min_variant_count
}
}

