% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-robust.R
\name{lapply_cond}
\alias{lapply_cond}
\title{Apply f to all conditions}
\usage{
lapply_cond(x, f, ...)
}
\arguments{
\item{x}{tibble of data per condition}

\item{f}{function to apply to each condition data}

\item{...}{parameters for f}
}
\value{
tibble
}
\description{
Wrapper for lapply
}
