% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-robust.R
\name{lapply_repl}
\alias{lapply_repl}
\title{Apply f to all replicates}
\usage{
lapply_repl(x, f, ..., cores = 1)
}
\arguments{
\item{x}{tibble of data per condition}

\item{f}{function to apply to each replicate data}

\item{...}{parameters for f}

\item{cores}{numer of cores to use}
}
\value{
tibble
}
\description{
Wrapper for lapply
}
