% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{learn_nmf}
\alias{learn_nmf}
\title{Estimate factorization rank from JACUSA2 output}
\usage{
learn_nmf(x, mods = NULL, nmf_args = NULL, nmf_seed = "nndsvd")
}
\arguments{
\item{x}{target data to fit, i.e. a matrix-like object - use `create_features()`}

\item{mods}{GRanges object of known modifications sites. Default: NULL.}

\item{nmf_args}{argument to be forward to `NMF::nmf`. @seealso NMF::nmf}

\item{nmf_seed}{Default: "nndsvd". @seealso NMF::nmfSeed}
}
\value{
Depends on `evaluate`: FALSE @seealso NMF::nmf, TRUE a list of objects.
}
\description{
Given a data matrix created by `create_data()` this function performs non-negative matrix factorization.
}
