% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-robust.R
\name{mapply_repl}
\alias{mapply_repl}
\title{Apply f to all replicates}
\usage{
mapply_repl(f, ..., MoreArgs = NULL)
}
\arguments{
\item{f}{function to apply to each replicate data}

\item{...}{see mapply}

\item{MoreArgs}{see mapply}
}
\value{
tibble
}
\description{
Wrapper for mapply.
}
