% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{mask_ref_base2bc}
\alias{mask_ref_base2bc}
\title{Mask a set of base changes.}
\usage{
mask_ref_base2bc(ref_base2bc, keep)
}
\arguments{
\item{ref_base2bc}{vector of base call changes.}

\item{keep}{vector of base call changes to be highlighted. All other will be renamed to "other".}
}
\value{
vector of base call changes.
}
\description{
When only a subset, here captured ny\code{keep}, of all possibles base changes is interesting, 
this function will hide the remaining base changes by renaming them to other.
}
