% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-sub.R
\name{mask_sub}
\alias{mask_sub}
\title{Mask a set of base substitutions.}
\usage{
mask_sub(subs, keep)
}
\arguments{
\item{subs}{vector of base call substitutions.}

\item{keep}{vector of base call substitutions to be highlighted. All other will be renamed to \emph{other}.}
}
\value{
vector of base call substitutions.
subs <- c("A->G", "A->C", "no change")
# "A->G" "other" "no change"
mask_sub(subs, c("A->G"))

# "other" "other" "no change"
mask_sub(subs, c("A->T"))
}
\description{
When only a subset of all possibles base substitutions is interesting, the remaining base substitutions can 
be masked. This function will hide the remaining base substitutions by renaming them to \emph{other}.
}
