% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{matrix2str}
\alias{matrix2str}
\title{Convert matrix to string representation.}
\usage{
matrix2str(matrix)
}
\arguments{
\item{matrix}{Matrix or list of matrices.}
}
\value{
Returns vector of Strings
}
\description{
Convert count matrix to vector of strings. Columns are separated by ",".
}
