% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-ratio.R
\name{max_observed_bc}
\alias{max_observed_bc}
\title{Calculates most frequent base calls}
\usage{
max_observed_bc(bases, ref = NULL)
}
\arguments{
\item{bases}{matrix of observed base call counts.}

\item{ref}{vector of reference bases. Default: NULL.}
}
\value{
vector of most frequent base call counts.
}
\description{
Calculates most frequent base call counts from a nx4 matrix \code{bases}.
Provide \code{ref} to ignore reference base call counts.
}
\examples{
bases <- matrix(c( 1,  0, 9, 0,
                   5,  0, 5, 0,
                   0, 10, 0, 0),
                ncol = 4, byrow = TRUE)
ref <- c("A", "A", "C")
# show most frequent base call counts
max_observed_bc(bases, ref)#'
}
