% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{merge_ref_base2bc}
\alias{merge_ref_base2bc}
\title{Merges base changes.}
\usage{
merge_ref_base2bc(ref_base2bc)
}
\arguments{
\item{ref_base2bc}{vector of base changes.}
}
\value{
merged base changes.
}
\description{
Merges basechanges. Usefull, when creating summaries for base changes with \code{add_summary()}.
Merging base changes with different reference bases is not allowed.
However, base changes with different non-reference bases can be merged, e.g.: c("A->G", "A->C")
will be merged to "A->CG"
}
