% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-sub.R
\name{merge_sub}
\alias{merge_sub}
\title{Merges base substitutions.}
\usage{
merge_sub(subs)
}
\arguments{
\item{subs}{vector of base substitutions.}
}
\value{
merged base substitutions.
}
\description{
Usefull, when creating summaries for base changes with \code{add_summary()}.
Merging base changes with different reference bases is not allowed.
However, base changes with different non-reference bases can be merged, e.g.: c("A->G", "A->C")
will be merged to "A->CG"
}
\examples{
sub <- c("A->G", "A->G")
# result: A->G
merge_sub(sub)
# result: A->CG
sub <- c("A->G", "A->C")
merge_sub(sub)
}
