% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modmap}
\alias{modmap}
\title{rRNA modification map of rRNAs according to Piekna-Przybylska2007}
\format{a data.frame with 3 elements:
\itemize{
  \item rrna: String, either "NR_003286_RNA18SN5" or "NR_003287_RNA28SN5".
  \item pos: Numeric 1-index position.
  \item base: String, reference base(A, C, G, T) or modification(Am, Cm, Gm, m1A, m3U, m5C, m62A, m6A, m7G, PseudoU, PseudoUm, or Um)
}}
\usage{
data(modmap)
}
\description{
Data (18S and 28S) has been download from: http://people.biochem.umass.edu/fournierlab/3dmodmap/
and coordinates habe been lifted to hg38.
}
\references{
D. Piekna-Przybylska, W. A. Decatur, and M. J. Fournier (2007).•↲
            The 3D rRNA modification maps database: with interactive tools for ribosome analysis.•↲
            Nucleic Acids Research.Jan;36:D178-83. Epub 2007 Oct 18.
}
\keyword{datasets}
