% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-ratio.R
\name{observed_bc}
\alias{observed_bc}
\title{Calculates observed base calls}
\usage{
observed_bc(bases)
}
\arguments{
\item{bases}{matrix of observed base call counts.}
}
\value{
vector of observed base calls.
}
\description{
Calculates observed base call counts stored in nx4 matrix \code{bases}.
}
\examples{
bases <- matrix(c( 1,  0, 9, 0,
                   5,  0, 5, 0,
                   0, 10, 0, 0),
                ncol = 4, byrow = TRUE)
# show observed base call counts
observed_bc(bases)
}
