% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Packs and converts a JACUSA2 result object to JACUSA2 result format. (EXPERIMENTAL)}
\usage{
pack(result)
}
\arguments{
\item{result}{object created by \code{read_result()}.}
}
\value{
result object in JACUSA2 result format.
}
\description{
Packs and converts a JACUSA2 result object to JACUSA2 result format.
This will be written to a file. Requires \code{result} to be a object 
created by \code{read_result()}. This functions converts from tidy data to 
JACUSA2 result format (wide-format). 
JACUSA2 specific attributes must be present in \code{result}.
}
