% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_base_ref2bc_ratio_matrix}
\alias{plot_base_ref2bc_ratio_matrix}
\title{Plot TODO}
\usage{
plot_base_ref2bc_ratio_matrix(result, ref_field, ref_base2bc = c())
}
\arguments{
\item{result}{object \code{read_results()}}

\item{ref_field}{TODO}

\item{ref_base2bc}{TODO}
}
\value{
Plot TODO
}
\description{
TODO
}
