% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{plot_bc_change_table}
\alias{plot_bc_change_table}
\title{Plot the distribution of base changes.}
\usage{
plot_bc_change_table(bc_change_tbl, bc_change = format_bc_change("A", "G"))
}
\arguments{
\item{bc_change_tbl}{Table object created by \code{table()}.}

\item{bc_change}{Base call change between condition 1 and 2 to score for.}
}
\description{
Plot the distribution of base changes.
}
