% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ecdf-allele-count.R
\name{plot_ecdf_allele_count}
\alias{plot_ecdf_allele_count}
\title{Plot cumulative allele count.}
\usage{
plot_ecdf_allele_count(
  result,
  data_desc,
  allele_count_column = "allele_count",
  ...
)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{data_desc}{vector of characters strings providing details about the sample.}

\item{allele_count_column}{character sting specifing the name of the column.}

\item{...}{parameters forwared to plot_ecdf_column.}
}
\value{
ggplot2 plot object.
}
\description{
Plots cumulative distribution of allele counts for single and multiple JACUSA2 result objects.
}
\seealso{
\code{plot_ecdf_column()}
}
