% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ecdf-arrest-rate.R
\name{plot_ecdf_arrest_rate}
\alias{plot_ecdf_arrest_rate}
\title{Plot cumulative arrest rate.}
\usage{
plot_ecdf_arrest_rate(result, data_desc,
  arrest_rate_column = "arrest_rate", ...)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{data_desc}{vector of characters strings providing details about the sample.}

\item{arrest_rate_column}{character sting specifing the name of the column.}

\item{...}{parameters forwarded to plot_ecdf_column.}
}
\value{
ggplot2 plot object.
}
\description{
Plots cumulative distribution of arrest rate for single and multiple JACUSA2 result objects.
}
