% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ecdf-column.R
\name{plot_ecdf_column}
\alias{plot_ecdf_column}
\title{Generic plot for cumulative data (EXPERIMENTAL)}
\usage{
plot_ecdf_column(result, column, data_desc, name = "Data description")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{column}{character string to be used for plotting.}

\item{data_desc}{vector of characters strings providing details about the sample.}

\item{name}{character string defines name of the legend.}
}
\value{
ggplot2 plot object.
}
\description{
Creates a generic plot of a \code{column} from a JACUSA2 \code{result} object and display  
verbose data descripton \code{data_desc}.
}
\details{
The user must make sure that dimensions (rows and length) of: 
\code{result} and \code{data_desc} match and that \code{column} exists 
in \code{result}. Furthermore, the user is responsible for correct filtering and grouping of the 
\code{result} object.
\code{data_desc} is used to provide a descriptive name for each observation.

TODO make column replicate optional
}
