% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf_coverage}
\alias{plot_ecdf_coverage}
\title{Plot cumulative read coverage}
\usage{
plot_ecdf_coverage(result)
}
\arguments{
\item{result}{object create by \code{read_result()}}
}
\value{
plot object
}
\description{
TODO
}
