% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ecdf-score.R
\name{plot_ecdf_score}
\alias{plot_ecdf_score}
\title{Plot cumulative score.}
\usage{
plot_ecdf_score(result, data_desc, column = "score", ...)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}}

\item{data_desc}{vector of strings providing details}

\item{column}{character string specifies column of score.}

\item{...}{parameters forwared to plot_ecdf_column.}
}
\value{
plot object
}
\description{
The user is responsible for correct filtering and grouping of the 
\code{result} object.
\code{data_desc} is used to provide a descriptive name for each observation.
}
