% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_meta_ecdf_column}
\alias{plot_meta_ecdf_column}
\title{Plot TODO}
\usage{
plot_meta_ecdf_column(js2, column)
}
\arguments{
\item{js2}{object \code{read_results()}}

\item{column}{to be used for plotting}
}
\value{
Plot TODO
}
\description{
TODO
}
