% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sub.R
\name{plot_sub}
\alias{plot_sub}
\title{Plot distribution of base substitutions.}
\usage{
plot_sub(result, column = "sub")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{column}{character string to be used for plotting.}
}
\value{
ggplot2 plot object.
}
\description{
Plots the distribution of base substitutions.
The user is responsible for correct filtering and grouping of the 
\code{result} object.
}
\examples{
library(magrittr)
data(rdd)
# 1st condition is gDNA - we use it to define ref_base
result <- redefine_ref_base(rdd, condition = 1)
result <- add_sub(rdd)
# filter by coverage and score
filtered <- result \%>\% 
  filter_by_coverage(min_coverage = 10) \%>\% 
  filter_by_min_score(min_score = 2)
plot_sub(filtered) + ggplot2::facet_grid(replicate ~ .)
}
