% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{predict_mods}
\alias{predict_mods}
\title{Predict modifications in JACUSA2 output}
\usage{
predict_mods(x, nmf_results = NULL)
}
\arguments{
\item{x}{target data to fit, i.e. a matrix-like object - use `create_data()`.}

\item{nmf_results}{Default: data(m6a_nmf_results).}
}
\value{
Score matrix.
}
\description{
Use data matrix and an matrix factorization to predict modification sites.
}
