% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdd}
\alias{rdd}
\title{Subset of RDDs detected by JACUSA2 in HEK-293 untreated cells}
\format{a tibble with 19 elements:
\itemize{
	\item contig: Character string representing the contig of the variant
	\item start: Numeric position of variant (>=0)
	\item end: Numeric corresponds to "start + 1"
	\item name: Character string. Currently, name of used method (\emph{call-\{1,2\}}, \emph{pileup}, \emph{rt-arrest}, or \emph{lrt-arrest})
	\item score: Numeric value representing the test-statistc. Higher values indicate more divergent pileups
	\item strand: Character representing strand information; "+", "-", or "."(no strand information available)
	\item info: Character string separated with ";" provding additional data for this specific site. Empty field is equal to "*"
	\item filter_info: ","-separated character string showing feature filter information. Empty field is equal to "*"
	\item ref_base: Character "A", "C", "G", "T", or "N" representing the reference base for this site - inverted when strand is "-".
	\item base_type: Character string inndicating what type of base calls are shown. Here: "total".
	\item condition: Numeric value: "1"(here: gDNA) or "2"(here: RNA)
	\item replicate: Numeric value: "1" for gDNA(no replicates) and "1" or "2" for RNA(1 replicate)
	\item bc_A: Numeric value representing base call count for A
	\item bc_C: Numeric value representing base call count for C
	\item bc_G: Numeric value representing base call count for G
	\item bc_T: Numeric value representing base call count for T
	\item bc: Observed base calls for this site. Combination of: "A", "C", "G", and "T"
   \item cov: Numeric value indicating the read coverage for this site
}}
\usage{
data(rdd)
}
\description{
Subset of RDDs detected by JACUSA2 in HEK-293 untreated cells
}
\references{
Piechotta, M.; Wyler, E.; Ohler, U.; Landthaler, M. & Dieterich, C.
JACUSA: site-specific identification of RNA editing events from replicate sequencing data 
BMC Bioinformatics, Springer Nature, 2017 , 18

A dataset containing a subset of RNA DNA differences (RDDs) identified by JACUSA2 in untreated HEK-293 cells. 
The fields are as follows:
}
\keyword{datasets}
