% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_jacusa}
\alias{read_jacusa}
\title{Read JACUSA ressult file}
\usage{
read_jacusa(file, ...)
}
\arguments{
\item{file}{String represents the filename of the JACUSA output.}

\item{...}{Additional parameters that will be forwarded to read.table.}
}
\description{
\code{read_jacusa()} Reads data that was generated by JACUSA.
}
