% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-result.R
\name{read_results}
\alias{read_results}
\title{Read multiple related JACUSA2 results}
\usage{
read_results(files, meta_conditions)
}
\arguments{
\item{files}{Vector of files}

\item{meta_conditions}{Vector of character vectors that correspond to files}
}
\value{
TODO
}
\description{
TODO
}
