% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-read.R
\name{read_str2read_matrix}
\alias{read_str2read_matrix}
\title{Convert read start, through, and end counts encoded as string vectors.}
\usage{
read_str2read_matrix(read_str, collapse = FALSE)
}
\arguments{
\item{read_str}{Vector of strings or list of string vectors that encode 
counts of read start, through, and end.}

\item{collapse}{Logical indicates if counts from replicates should be aggregated.}
}
\value{
Returns a matrix or a list of matrices of read start, through, and 
end counts.
}
\description{
\code{read_str2read_matrix()} converts read start, through ,and end counts 
string vectors to count matrices.
}
