% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine-ref.R
\name{redefine_ref}
\alias{redefine_ref}
\title{Redefine reference base with observed base calls.}
\usage{
redefine_ref(result, ref_cond, base_type = "bases", keep_old = NULL)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{ref_cond}{numeric specifying the condition to define ref.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}

\item{keep_old}{string specifying column to store old ref. Default: NULL.}
}
\value{
result object with base changes added.
}
\description{
There must be only one reference base. A->G is okay, but AG->G is NOT allowed!
}
