% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine-ref-base.R
\name{redefine_ref_base}
\alias{redefine_ref_base}
\title{Redefine reference base with observed base calls.}
\usage{
redefine_ref_base(result, ref_condition, base_type = "bases",
  keep_old = NULL)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{ref_condition}{TODO}

\item{base_type}{TODO}

\item{keep_old}{string TODO Default: NULL}
}
\value{
result object with base changes added.
}
\description{
There must be only one reference base. A->G is okay, but AG->G is NOT allowed!
}
