% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{reduce_coef}
\alias{reduce_coef}
\title{Reduce coefficient matrix}
\usage{
reduce_coef(nmf_results, meta_conds)
}
\arguments{
\item{nmf_results}{object created by `learn_nmf()`.}

\item{meta_conds}{names of the experiments in nmf_results}
}
\value{
nmf_results with modified coefficient matrix (t_h).
}
\description{
Reduce coefficient matrix to make predictions on one experiment data.
}
