% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnam}
\alias{rnam}
\title{rRNA modification map of rRNAs according to Taika et al.}
\format{
a data.frame with 3 elements:
\itemize{
  \item rrna: String, either "18S", "28S", or "5.8S"
  \item pos: Numeric 1-index position.
  \item base: String, reference base(A, C, G, T) or modification(ac4C, AM, Cm, 
        Gm, m1A, m1acp3pusU, m3U, m5C, m62A, m6A, m7G, ND, psU, psUm, Um)
}
}
\usage{
data(rnam)
}
\description{
Data (18S, 28S, and 5.8S) has been extracted from retrieved from PMID: 30202881.
}
\references{
Landscape of the complete RNA chemical modifications in the human 80S ribosome
            Masato Taoka, Yuko Nobe, Yuka Yamaki, Ko Sato, Hideaki Ishikawa, 
            Keiichi Izumikawa, Yoshio Yamauchi, Kouji Hirota, Hiroshi Nakayama, 
            Nobuhiro Takahashi et al. 
            Nucleic Acids Research, Volume 46, Issue 18, 12 October 2018, Pages 9289–9298, 
            https://doi.org/10.1093/nar/gky811
}
\keyword{datasets}
