% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-robust.R
\name{robust}
\alias{robust}
\title{Check if data is robust}
\usage{
robust(data)
}
\arguments{
\item{data}{tibble of numeric.}
}
\value{
logical vector indicating robust rows.
}
\description{
Tests if all values that have been observed (>0) in all conditions (op = |)
Supports 1 or 2 conditions.
}
