% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-robust-variants.R
\name{robust_variants}
\alias{robust_variants}
\title{TODO}
\usage{
robust_variants(condition, bases)
}
\arguments{
\item{condition}{TODO}

\item{bases}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}
