% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot to disc}
\usage{
save_plot(filename, plot, ...)
}
\arguments{
\item{filename}{character vector TODO}

\item{plot}{object to be plotted TODO}
}
\description{
TODO
}
