% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{score_bc_change}
\alias{score_bc_change}
\title{Calculates the fraction of editing sites among RDDs in JACUSA result file.}
\usage{
score_bc_change(bc_change_tbl, bc_change = format_bc_change("A", "G"))
}
\arguments{
\item{bc_change_tbl}{Vector object created by \code{table()}.}

\item{bc_change}{String that identifies true editing, e.g.: "A->G". Use: \code{format_bc_change()}}
}
\value{
Returns a numeric value that represent the fraction of editing sites among RDDs.
}
\description{
\code{score_bc_change()} calculates the fraction of editing sites among RDDs in JACUSA result file.
}
\examples{
## add fields automatically
jacusa <- add_bc_change(rdd, aa = TRUE)
bc_change_tbl <- table(jacusa$bc_change)
score_bc_change(bc_change_tbl)

}
