% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-ratio.R
\name{sub_ratio}
\alias{sub_ratio}
\title{Calculates base substitution ratio (e.g.: editing frequency).}
\usage{
sub_ratio(ref, bases, bc = NULL)
}
\arguments{
\item{ref}{vector of reference bases.}

\item{bases}{matrix of observed base call counts.}

\item{bc}{vector strings that represent observed base calls. Default: NULL.}
}
\value{
vector of base call substitution ratios.
}
\description{
Calculates base substitution ratio (e.g.: editing frequency) for \code{ref} and base counts stored
in nx4 matrix \code{bases}.
\code{ref} must have only one reference base. A->G is okay, but AG->G is NOT allowed!
If \code{bc} is not provided, count the substitution ratio of the largest non-reference base
(@seealso \code{max_observed_bc}). Make sure that number of alleles in ref(erence) and bases
is 2 at maximum.
}
\examples{
ref <- c("A", "A", "T")
# Only one non-ref. base
bases <- matrix(c( 1,  0, 9, 0,
                   5,  0, 5, 0,
                   0, 10, 0, 0),
                ncol = 4, byrow = TRUE)
# show base substitution ratios for observed base calls
sub_ratio(ref, bases)

# > one non-ref. base
bases <- matrix(c( 1,  2, 9, 1,
                   5,  3, 5, 1,
                   2, 10, 2, 1),
               ncol = 4, byrow = TRUE)
# we are only interested in A->G, T->C
bc <- c("G", "G", "C")
# show base substitution ratios for A->G, T->C
sub_ratio(ref, bases, bc)
}
