% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-ratio.R
\name{sub_ratio_col}
\alias{sub_ratio_col}
\title{Column name for base substitution ratio for \code{base_type}.}
\usage{
sub_ratio_col(base_type = "bases")
}
\arguments{
\item{base_type}{string defining the column to use as base counts. Default: bases.}
}
\value{
string the represents the column name for base substition ratio for \code{base_type}.
}
\description{
Column name for base substitution ratio for \code{base_type}. 
\code{base_type} will be used to create the column name:
\code{sub[_base_type]}.
}
