% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{through_col}
\alias{through_col}
\title{Column name for through base counts for \code{suffix}.}
\usage{
through_col(suffix = NULL)
}
\arguments{
\item{suffix}{string to specifiy base columns to retrieve through bases Default: NULL}
}
\value{
string the represents the column name for through bases marked with \code{suffix}.
}
\description{
Column name for through base counts for \code{suffix}. 
\code{suffix} will be used to create the column name:
\code{through[_suffix]}".
}
