% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-ratio.R
\name{variant_bc}
\alias{variant_bc}
\title{Calculates non-reference base calls}
\usage{
variant_bc(ref, bc)
}
\arguments{
\item{ref}{vector of reference bases..}

\item{bc}{vector of base calls.}
}
\value{
vector of most frequent base call counts.
}
\description{
Calculates non-reference base call counts from a vector \code{bc} and 
\code{ref}(ference) bases. Make sure the combination of base calls and 
reference bases <= 2 alleles.
}
\examples{
ref <- c("A", "A")
bc <- c("A", "AG")
# show non-reference base calls
variant_bc(ref, bc)
}
