% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-write.R
\name{write_bedGraph}
\alias{write_bedGraph}
\title{Write vector of values as bedGraph file}
\usage{
write_bedGraph(file, contig, start, end, value)
}
\arguments{
\item{file}{String represents the filename of the BED graph file.}

\item{contig}{Vector of character vectors}

\item{start}{Vector of numericals zero indexed (>= 0)}

\item{end}{Vector of numericals half opened: [start, end)}

\item{value}{Vector of characters or numericals the reference value(s)}
}
\description{
\code{write_bedGraph()} Writes a bedGraph conform file with coordinates
defined by vectors: "contig", "start", and "end". According to:
https://genome.ucsc.edu/goldenPath/help/bedgraph.html
"start" and "end" are zero index and half opened. The caller needs to make sure 
that this is fullfilled.
}
