% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-result.R
\name{write_result}
\alias{write_result}
\title{Writes a JACUSA2 result object to a file.}
\usage{
write_result(result, file, txt = "")
}
\arguments{
\item{result}{object create by \code{read_result}.}

\item{file}{character string naming a file for writing}

\item{txt}{character string to be added to the JACUSA2 header. If "", nothing will be added.}
}
\description{
Writes a JACUSA2 result object to a file. 
The optional character \code{txt} will be added to the JACUSA2 header 
along with the version of this package.
}
