% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{add_arrest_rate}
\alias{add_arrest_rate}
\title{Add arrest rate.}
\usage{
add_arrest_rate(result, suffix = NULL)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{suffix}{TODO Default: NULL}
}
\value{
result object with arrest rate field "arrest_rate[_\code{suffix}]" added.
}
\description{
Adds arrest rate calculated from \code{arrest[_suffix]} and \code{through[_suffix]} columns.
}
