% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{add_arrest_rate}
\alias{add_arrest_rate}
\title{Add arrest rate.}
\usage{
add_arrest_rate(result)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}
}
\value{
result object with arrest rate field \code{arrest_rate} added.
}
\description{
Calculates and adds arrest rate from \code{arrest} and \code{through} columns.
}
