% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{add_arrest_rate}
\alias{add_arrest_rate}
\title{Add arrest rate.}
\usage{
add_arrest_rate(result, suffix = NULL)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{suffix}{string added to specifiy base columns to calculate arrest rate. Default: NULL}
}
\value{
result object with arrest rate field \code{arrest_rate[_suffix]} added.
}
\description{
Adds arrest rate calculated from \code{arrest[_suffix]} and \code{through[_suffix]} columns.
}
\examples{
data(HIVRT)
result <- add_arrest_rate(HIVRT)
}
