% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_bc}
\alias{add_bc}
\title{Add string vector of observed base calls for each condition to JACUSA list 
object.}
\usage{
add_bc(jacusa, bc_matrix1 = jacusa$bc_matrix1,
  bc_matrix2 = jacusa$bc_matrix2, aa = FALSE)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{bc_matrix1}{Base call matrix of list of base call matrices for 
condition 1.}

\item{bc_matrix2}{Base call matrix of list of base call matrices for 
condition 2.}

\item{aa}{Logical indicates if missing fields should added automatically.}
}
\value{
Returns a JACUSA list object with additional Base call (BC) fields: 
"bc1" and "bc2".
}
\description{
\code{add_bc()} calculates a vector of observed base calls for each condition 
and adds "bc1" and "bc2" fields to the JACUSA list object. 
Make sure \code{add_bc_matrix()} has been called to populate "bc_matrix1" 
and "bc_matrix2" fields or set option "aa" (i.e. auto_add) to TRUE.
All necessary but missing fields will be added automatically, possibly 
overriden existing fields of the JACUSA list object.
}
\examples{
## add fields manually
jacusa <- add_bc_matrix(rdd)
jacusa <- add_bc(jacusa)
## plot distribution of base calls in condition1
barplot(table(jacusa$bc1))

## add missing fields automatically
jacusa <- add_bc(jacusa, aa = TRUE)
barplot(table(jacusa$bc1))

}
