% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_bc_change}
\alias{add_bc_change}
\title{Add base call change for RDD comparisons of a JACUSA result file.}
\usage{
add_bc_change(jacusa, bc1 = jacusa$bc1, bc2 = jacusa$bc2, aa = FALSE)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}

\item{bc1}{Vector of observed base calls for condition 1.}

\item{bc2}{Vector of observed base calls for condition 2.}

\item{aa}{Logical indicates if missing fields should added automatically.}
}
\value{
Returns a JACUSA object list with the additional "bc_change" field.
}
\description{
\code{add_bc_change()} adds base call changes for gDNA vs. cDNA comparisons.
Make sure \code{add_bc()} has been called to populate "bc1" and "bc2" fields or 
set option "aa" (i.e. auto_add) to TRUE.
All necessary but missing fields will be added automatically, possibly 
overriden existing fields of the JACUSA list object.
}
\examples{
## populate necessary fields manually
jacusa <- add_bc_matrix(rdd)
jacusa <- add_bc(jacusa)
jacusa <- add_bc_change(jacusa)
## plot distribution of base changes
barplot(table(jacusa$bc_change))

## populate necessary fields automatically
jacusa <- add_bc_change(jacusa, aa = TRUE)
## plot distribution of base changes
barplot(table(jacusa$bc_change))

}
