% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Add coverage field to JACUSA2 object.}
\usage{
add_coverage(jacusa2)
}
\arguments{
\item{jacusa2}{object created by \code{read_result()}.}
}
\value{
Returns the initial object with coverage key added.
}
\description{
\code{add_coverage()} calculates and adds read coverage.
}
