% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Add coverage field to JACUSA2 result object.}
\usage{
add_coverage(result)
}
\arguments{
\item{result}{object created by \code{read_result()}.}
}
\value{
result object with coverage field added.
}
\description{
\code{add_coverage()} calculates and adds read coverage.
}
