% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Add coverage to JACUSA2 result object}
\usage{
add_coverage(result)
}
\arguments{
\item{result}{created by \code{read_result()}}
}
\value{
result object with coverage added
}
\description{
Calculates and adds read coverage.
}
