% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_coverage}
\alias{add_coverage}
\title{Add coverage fields to JACUSA list object.}
\usage{
add_coverage(jacusa)
}
\arguments{
\item{jacusa}{List object created by \code{read_jacusa()}.}
}
\value{
Returns the initial list with coverage fields added.
}
\description{
\code{add_coverage()} calculates and adds read coverage to list of sites. 
This function will add 5 fields to the initial list: 
cov1, cov2: total read coverage per sample,
covs1, covs2: read coverage per sample and replicate, and
cov: total read coverage (cov1 + cov2).
}
\examples{
## add coverage  to data
jacusa <- add_coverage(rdd)
## filter/keep sites that have a total coverage of 30
jacusa <- filter_coverage(jacusa, "cov", 30)

}
