% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data-desc.R
\name{add_data_desc}
\alias{add_data_desc}
\title{Adds data description to JACUSA2 result object}
\usage{
add_data_desc(result, condition_desc = c(),
  format = "\%CONDITION\%_\%REPLICATE\%")
}
\arguments{
\item{result}{object created by \code{read_result()}}

\item{condition_desc}{TODO list of vector mapping condition to more verbose description}

\item{format}{character vector must contain place holder: \%CONDITION\% and \%REPLICATE\%}
}
\value{
result object with data description added
}
\description{
Adds a more descriptive name to the resut object, beautifying condition and replicate information.
}
