% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-non-ref-base-ratio.R
\name{add_non_ref_base_ratio}
\alias{add_non_ref_base_ratio}
\title{Adds non reference base ratio to JACUSA2 result object.}
\usage{
add_non_ref_base_ratio(result)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}
}
\value{
result object with base changes added.
}
\description{
Total read coverage = reference BC + non-reference base BCs.
}
