% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-ref-base2bc.R
\name{add_ref_base2bc}
\alias{add_ref_base2bc}
\title{Adds base change column to JACUSA2 result object.}
\usage{
add_ref_base2bc(result, ref_field)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{ref_field}{character string "ref_base" or a condition number (1 or 2) to define reference base.}
}
\value{
result object with base changes added.
}
\description{
The reference base can be defined either by providing the column name "ref_base" 
or an integer that specifies the condition. 
When comparing DNA vs. RNA  sequencing samples, base calls identified in DNA might differ from
the reference base provided by "ref_base" column. 
There must be only one reference base. A->G is okay, but AG->G is NOT allowed! 
Make sure to filter \code{result} with \code{filter_by_alleles()} to comply with this restriction.
}
