% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_ref_base2bc_ratio}
\alias{add_ref_base2bc_ratio}
\title{Add reference base to base call change ratio (e.g.: editing frequency) to a JACUSA2 object.}
\usage{
add_ref_base2bc_ratio(result, ref_field)
}
\arguments{
\item{result}{object created by \code{read_result()}.}

\item{ref_field}{String ("ref_base" for 1 condition experiment) or integer of DNA condition}
}
\value{
result object with base change and base change ratio added
}
\description{
\code{add_ref_base2bc_ratio()} adds base call change ratio (e.g.: editing frequency).
Adds key "ref_base2bc" key to JACUSA2 object.
}
