% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-bc.R
\name{add_ref_base2bc_ratio}
\alias{add_ref_base2bc_ratio}
\title{Add reference base to base call change ratio (e.g.: editing frequency) to a JACUSA2 object.}
\usage{
add_ref_base2bc_ratio(jacusa2, ref_field = "ref_base")
}
\arguments{
\item{jacusa2}{object created by \code{read_result()}.}

\item{ref_field}{String or integer}
}
\value{
Returns a JACUSA2 object
}
\description{
\code{add_ref_base2bc_ratio()} adds base call change ratio (e.g.: editing frequency).
}
