% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{arrest_col}
\alias{arrest_col}
\title{Column name for arrest base counts for \code{suffix}.}
\usage{
arrest_col(suffix = NULL)
}
\arguments{
\item{suffix}{string to specifiy base columns to retrieve arrest bases Default: NULL}
}
\value{
string the represents the column name for arrest bases marked with \code{suffix}.
}
\description{
Column name for arrest base counts for \code{suffix}. 
\code{suffix} will be used to create the column name:
\code{arrest[_suffix]}".
}
