% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{arrest_rate}
\alias{arrest_rate}
\title{Calculate arrest rate.}
\usage{
arrest_rate(arrest_cov, through_cov)
}
\arguments{
\item{arrest_cov}{numeric vector of read coverage that exhibit read arrest.}

\item{through_cov}{numeric vector of read coverage that DO NOT exhibit read arrest.}
}
\value{
numeric vector of arrest rates.
}
\description{
Calculates arrest rate from arrest and through reads.
}
\examples{
arrest_cov <- c(10, 1, 0)
through_cov <- c(0, 1, 1)
str(arrest_rate(arrest_cov, through_cov))
}
