% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{arrest_rate_col}
\alias{arrest_rate_col}
\title{Column name for arrest rate for \code{suffix}.}
\usage{
arrest_rate_col(suffix = NULL)
}
\arguments{
\item{suffix}{TODO}
}
\value{
string the represents the column name for arrest rate for \code{suffix}.
}
\description{
Column name for arrest ratio for \code{suffix}. 
\code{suffix} will be used a suffix to create the column name:
"arrest_rate[_\code{suffix}]".
}
