% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-count.R
\name{base_count}
\alias{base_count}
\title{Retrieves the total number of observed unique bases.}
\usage{
base_count(bc_obs, ref_base = NULL)
}
\arguments{
\item{bc_obs}{vector of strings with observed base calls.}

\item{ref_base}{vector of strings with reference base,}
}
\value{
numeric total number of bases.
}
\description{
Retrieves the total number of observed unique bases for \code{base_type}.
If \code{ref_base} is provided, reference base will be considered for counting, e.g.:
At some site only base 'A' was observed but the reference was 'T'.
\code{base_count} will be:
\itemize{
 \item 1, when \code{ref_base} is NOT provided, or
 \item 2, when \code{ref_base} is provided.
}
}
\examples{
bc_obs <- c("A", "AG", "C")
str(base_count(bc_obs))

# use default observed base calls and condsider reference base
ref_base <- c("A", "A", "T")
str(base_count(bc_obs, ref_base))
}
