% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sub.R
\name{base_sub}
\alias{base_sub}
\title{Calculates reference base to base call (base substitution).}
\usage{
base_sub(ref, bases)
}
\arguments{
\item{ref}{vector of reference bases.}

\item{bases}{vector of base calls or tibble of base call counts.}
}
\value{
vector of formatted base call substitutions.
}
\description{
Calculates and formats base changes for \code{ref} and \code{bc}.
All elements in \code{ref} must be all length 1.
}
\examples{
ref <- c("A", "A", "C", "A")
bases <- c("AG", "G", "C", "G")
base_sub(ref, bases)
}
