% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-sub.R
\name{base_sub}
\alias{base_sub}
\title{Calculates reference base to base call (base substitution).}
\usage{
base_sub(ref_base, bc)
}
\arguments{
\item{ref_base}{vector of reference bases.}

\item{bc}{vector of base calls.}
}
\value{
vector of formatted base call substitutions.
}
\description{
Calculates and formats base changes for \code{ref_base} and \code{observed_bc}.
All elements in \code{ref_bases} must be all length 1.
}
\examples{
ref_base <- c("A", "A", "C", "A")
bc <- c("AG", "G", "C", "G")
base_sub(ref_base, bc)
}
