% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-max-alleles.R
\name{check_max_alleles}
\alias{check_max_alleles}
\title{Check if site contains the max number of allowed alleles}
\usage{
check_max_alleles(result, max_alleles = 2, use_ref_base = TRUE)
}
\arguments{
\item{result}{object created by \code{read_result()}.}

\item{max_alleles}{Integer number of maximal allowed alleles per site.}

\item{use_ref_base}{boolean indicating if ref_base should used to count alleles per site (Default: TRUE).}
}
\value{
boolean
}
\description{
Return TRUE|FALSE if the JACUSA2 result object has the maximum number of allowed 
alleles per site.
}
