% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_max_alleles}
\alias{check_max_alleles}
\title{Check if site has the max number of allowed alleles.}
\usage{
check_max_alleles(jacusa2, max_alleles = 2)
}
\arguments{
\item{jacusa2}{object created by \code{read_result()}.}

\item{max_alleles}{Integer number of maximal allowed alleles per site.}
}
\value{
TRUE|FALSE
}
\description{
Return TRUE|FALSE if the JACUSA2 object has the maximum number of allowed 
alleles per site.
}
