% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{coverage}
\alias{coverage}
\title{Calculate coverage for structured column.}
\usage{
coverage(bases, cores = 1)
}
\arguments{
\item{bases}{structured column of bases}

\item{cores}{number of cores to use}
}
\value{
structure column with coverages
}
\description{
Calculate coverage for structured column.
}
