% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Calculate coverage from base counts.}
\usage{
coverage(bases)
}
\arguments{
\item{bases}{nx4 matrix or data frame of base call counts.}
}
\value{
numeric vector of coverage for \code{bases}.
}
\description{
Calculates coverage from base counts. 
\code{bases} is expected to be a nx4 matrix or data frame.
}
\examples{
bases <- matrix(
  c( 1, 0, 0, 0,
    10, 0, 0, 0,
     6, 2, 0, 0),
  ncol = 4, 
  byrow = TRUE
)
str(coverage(bases))
}
