% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{expand_tag}
\alias{expand_tag}
\title{Expand tagged reads}
\usage{
expand_tag(result, cores = 1)
}
\arguments{
\item{result}{object tagged and not tagged structured base columns.}

\item{cores}{Integer defines how many cores to use.}
}
\description{
TODO
}
