% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{filter_artefact}
\alias{filter_artefact}
\title{Filters sites by artefacts.}
\usage{
filter_artefact(filter, artefacts)
}
\arguments{
\item{filter}{vector of strings that contains artefact filter information.}

\item{artefacts}{vector of characters that correspond to feature/artefact filters to be filtered.}
}
\value{
vector of logical.
}
\description{
Removes sites that sites that have been marked by feature/artefact filter.
}
\examples{
data(rdd)
# remove sites that are marked by artefact filter "D"
grouped <- group_by_site(rdd)
filtered <- filter_by(grouped, filter_artefact(filter, c("D")))
str(filtered)
}
