% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-filter.R
\name{filter_artefact}
\alias{filter_artefact}
\title{Filters sites by artefacts.}
\usage{
filter_artefact(filter_info, artefacts)
}
\arguments{
\item{filter_info}{vector of strings that contains artefacts filter information.}

\item{artefacts}{vector of characters that correspond to feature/artefact filters.}
}
\value{
vector of logical.
}
\description{
Removes sites that sites that have been marked by feature/artefact filter.
}
\examples{
data(rdd)
# remove sites that are marked by artefact filter "D"
filtered <- filter_by(resultm, filter_artefact(result, c("D")))
str(filtered)
}
