% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_alleles}
\alias{filter_by_alleles}
\title{Filter sites by the number of alleles per site.}
\usage{
filter_by_alleles(jacusa2, max_alleles = 2, ref_base = TRUE)
}
\arguments{
\item{jacusa2}{object created by \code{read_result()}.}

\item{max_alleles}{integer number of maximal allowed alleles per site.}

\item{ref_base}{boolean indicating ref_base should used to count alleles per site (Default: TRUE).}
}
\value{
Returns JACUSA2 object with sites where max_alleles is obeyed.
}
\description{
Removes sites that have too many alleles.
}
