% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-score.R
\name{filter_by_max_score}
\alias{filter_by_max_score}
\title{Restain sites with maximal score.}
\usage{
filter_by_max_score(result, max_score, field = "score")
}
\arguments{
\item{result}{object created by \code{read_result()}.}

\item{max_score}{Numeric value that represents the maximal score.}

\item{field}{character vector that identifies the column}
}
\value{
JACUSA2 result object with sites with score <= max_score.
}
\description{
Retains sites that have score <= max_score.
}
