% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-score.R
\name{filter_by_min_score}
\alias{filter_by_min_score}
\title{Retains sites by minimal score.}
\usage{
filter_by_min_score(result, min_score, column = "score")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{min_score}{numeric value that specifies the minimal score.}

\item{column}{character string that specifies the column of the score.}
}
\value{
result object with sites where score >= min_score.
}
\description{
Retains sites that have score >= min_score.
}
