% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-score.R
\name{filter_by_min_score}
\alias{filter_by_min_score}
\title{Restain sites by minimal score}
\usage{
filter_by_min_score(result, min_score, field = "score")
}
\arguments{
\item{result}{object created by \code{read_result()}.}

\item{min_score}{Numeric value that represents the minimal score.}

\item{field}{character vector that identifies the column}
}
\value{
JACUSA2 result object with sites with score >= min_score
}
\description{
Retains sites that have score >= min_score.
}
