% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-robust-variants.R
\name{filter_by_robust_variants}
\alias{filter_by_robust_variants}
\title{Retains sites that contain a variant base in all replicates in at least one condition.}
\usage{
filter_by_robust_variants(result, base_type = "bases", ...)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining the column to use as base counts. Default: bases.}

\item{...}{passed to internal \code{group_by_site} statement. Add additional grouping variable here, e.g.: "meta_condition".}
}
\value{
result object with sites where at least one condition contains the variant base in all replicates.
}
\description{
Enforces that at least one condition contains the variant base in all replicates.
}
\examples{
data(rdd)
str(filter_by_robust_variants(rdd))
}
