% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-robust-varianst.R
\name{filter_by_robust_variants}
\alias{filter_by_robust_variants}
\title{Retains sites that contain a variant base in all replicates in at least one condition}
\usage{
filter_by_robust_variants(result)
}
\arguments{
\item{result}{object created by \code{read_result()}.}
}
\value{
Returns JACUSA2 result object with sites where at least one condition contains the variant base in all replicates.
}
\description{
Enforces that at least one condition contains the variant base in all replicates.
}
