% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{format_bc_change}
\alias{format_bc_change}
\title{Format base change / base substitution.}
\usage{
format_bc_change(base1, base2, sep = BC_CHANGE_SEP,
  no_change = BC_CHANGE_NO_CHANGE)
}
\arguments{
\item{base1}{vector of bases: reference}

\item{base2}{vector of bases: observed bases.}

\item{sep}{string: "base1"sep"base2".}

\item{no_change}{character vector: how to format no changes.}
}
\value{
Vector of base changes.
}
\description{
Formats base change (e.g.: RNA editing) for two base vectors.
}
