% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{format_bc_change}
\alias{format_bc_change}
\title{Format base call change between two conditions.}
\usage{
format_bc_change(bc1, bc2, sep = .BC_CHANGE_SEP)
}
\arguments{
\item{bc1}{Vector of base calls for condition 1.}

\item{bc2}{Vector of base calls for condition 2.}

\item{sep}{String: "bc1"sep"bc2".}
}
\value{
Vector of base call changes between condition 1 and 2.
}
\description{
Formats base call change (e.g.: RNA editing) for two base call vectors.
}
