% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gather_repl}
\alias{gather_repl}
\title{Extract values from a structured column.}
\usage{
gather_repl(result, col)
}
\arguments{
\item{result}{JACUSA2 result object}

\item{col}{name of structured column}
}
\value{
extracted column
}
\description{
Extract valeus from a structured column and adds "contig:start-end:strand".
This is usefull if data should be used with `ggplot2`.
}
