% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_allele_count}
\alias{get_allele_count}
\title{Calculate allele count per site.}
\usage{
get_allele_count(bc_matrix1, bc_matrix2, max_alleles = 2)
}
\arguments{
\item{bc_matrix1}{Base call matrix or list of base call matrices for condition 1.}

\item{bc_matrix2}{Base call matrix or list of base call matrices for condition 2.}

\item{max_alleles}{Integer number of maximal alleles per site.}
}
\value{
Returs JACUSA list object with sites where max_alleles is obeyed.
}
\description{
Calculate allele count per site.
}
