% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{get_arrest_rate}
\alias{get_arrest_rate}
\title{Retrieve arrest rate.}
\usage{
get_arrest_rate(result, suffix = NULL)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{suffix}{string added to specifiy base columns to retrieve arrest rate. Default: NULL}
}
\value{
numeric vector with arrest rates for base counts with \code{suffix}.
}
\description{
Retrieves arrest rate from \code{arrest[_suffix]} and \code{through[_suffix]} columns.
}
\examples{
data(HIVRT)
str(get_arrest_rate(HIVRT))
}
