% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_bc}
\alias{get_bc}
\title{Calculate observed base calls from base call matrix or a list of base call matrices.}
\usage{
get_bc(bc_matrix)
}
\arguments{
\item{bc_matrix}{Base call matrix or list of bc matrices.}
}
\value{
Returns vector of strings of observed base calls in m.
}
\description{
Calculate observed base calls from base call matrix or a list of base call matrices.
}
