% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_bc_change}
\alias{get_bc_change}
\title{Calculate base call change for RDD comparisons.}
\usage{
get_bc_change(ref_base, observed_bc, sep = .BC_CHANGE_SEP)
}
\arguments{
\item{ref_base}{Vector of reference bases.}

\item{observed_bc}{Vector of base calls.}

\item{sep}{String: "ref_base"sep"observed_bcs".}
}
\value{
Vector of base call changes.
}
\description{
TODO
}
