% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_bc_change}
\alias{get_bc_change}
\title{Calculate base call change for RDD comparisons of JACUSA result file.}
\usage{
get_bc_change(bc1, bc2, sep = .BC_CHANGE_SEP)
}
\arguments{
\item{bc1}{Vector of base calls for condition 1.}

\item{bc2}{Vector of base calls for condition 2.}

\item{sep}{String: "bc1"sep"bc2".}
}
\value{
Vector of base call changes for condition 1 and 2.
}
\description{
In the context of RDD detection, it is assumed that condition 1 represents the DNA and 
condition 2 the RNA.
}
