% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bc.R
\name{get_bc_change_ratio}
\alias{get_bc_change_ratio}
\title{Calculates base call change ratio (e.g.: editing frequency) for RDDs in a JACUSA result file.}
\usage{
get_bc_change_ratio(bc1, bc2, bc_matrix2)
}
\arguments{
\item{bc1}{Vector of base calls for condition 1.}

\item{bc2}{Vector of base calls for condition 2.}

\item{bc_matrix2}{Base call matrix or list of base call matrices.}
}
\value{
Returns vector of base call change ratios average and against all 
replicates of condition 2.
}
\description{
Calculates base call change ratio (e.g.: editing frequency) for RDDs in a JACUSA result file.
}
