% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{get_cov}
\alias{get_cov}
\title{Retrieve coverage for \code{base_type}.}
\usage{
get_cov(result, base_type = "bases")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{base_type}{string defining base counts that were used to define base substitution. Default: bases.}
}
\value{
numeric vector with coverage information for \code{base_type}.
}
\description{
Retrieves coverge for base counts stored under "\code{base_type}".
}
\examples{
TODO
}
