% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_info_keys}
\alias{get_info_keys}
\title{Get keys of info field.}
\usage{
get_info_keys(info)
}
\arguments{
\item{info}{Vector of strings.}
}
\value{
returndfs a vector of available keys.
}
\description{
Get keys of info field.
}
