% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_primary_sites}
\alias{get_primary_sites}
\title{Get number of primary sites}
\usage{
get_primary_sites(result)
}
\arguments{
\item{result}{object created by \code{read_result()}}
}
\value{
intger the numner of primary sites
}
\description{
Retrieves the number of primary sites - unsplit and unique coordinates (TODO).
}
