% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{get_ref_base2bc_ratio}
\alias{get_ref_base2bc_ratio}
\title{Calculates base call change ratio (e.g.: editing frequency).}
\usage{
get_ref_base2bc_ratio(ref_base, bc_matrix, observed_bc = NULL)
}
\arguments{
\item{ref_base}{vector of reference bases.}

\item{bc_matrix}{matrix of observed base calls.}

\item{variant_bc}{TODO}
}
\value{
vector of base call change ratios.
}
\description{
Calculates base call change ratio (e.g.: editing frequency).
The following restrictions apply to column "ref_base" and "bc_matrix" apply:
\itemize{
  \item In "ref_base" must have only one reference base. A->G is okay, but AG->G is NOT allowed!
  \item In "bc_matrix" must contain only one non-reference base. A->G is okay, but A->CG is NOT allowed!
}
Make sure to filter \code{result} with \code{filter_by_alleles()} to comply with this restrictions.
}
