% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by-site.R
\name{group_by_site}
\alias{group_by_site}
\title{Groups JACUSA2 result object by coordinates and optional columns.}
\usage{
group_by_site(result, ...)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{...}{forward options to dplyr::group_by.}
}
\value{
grouped result object.
}
\description{
This is supposed to be used within a dplyr pipeline to
group JACUSA2 result object by coordinates: contig, start, end, strand -> a site. 
This allows to correctly group in case of structured data that was generated with JACUSA2: 
\emph{rt-arrest}, \emph{lrt-arrest}, and read/base stratification.
}
