% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by-site.R
\name{group_by_site}
\alias{group_by_site}
\title{TODO}
\usage{
group_by_site(result, opt_vars = c(), ...)
}
\arguments{
\item{result}{object created by \code{read_result(s)()}.}

\item{opt_vars}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
