% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{package}
\name{JACUSA2helper}
\alias{JACUSA2helper}
\alias{JACUSA2helper-package}
\title{JACUSA2helper: A package for post-processing JACUSA (1.x and 2.x) result files.}
\description{
The JACUSA2helper package provides the following categories of important functions:
input/output, get, add, and filter to post-process result files of JACUSA (1.x and 2.x).
}
\details{
\itemize{
 \item[call{1,2}]calling variants from one condition or condition comparisons.
 \item[pileup] SAMtools like pileup
 \item[rt-arrest] identification of RNA modifications by reverse transcription arrest.
 \item[lrt-arrest] combination of varidant discovery and read arrest sites. 
}

One major new feature of JACUSA 2.x is the identificaion of RNA modifications by reverse
transcription arrest.
tidy data, dplyr

TODO When calling RDDs where the RNA-Seq sample has been generated with stranded sequencing library, 
base change(s) can be directly inferred and if necessary base calls can be inverted.

TODO Warning: Some function do not support replicates or are exclusively applicable on RDD or RRD 
result files!
use jacusa2 <- read_result("jacusa2.out") to read JACUSA2 output

The central object in Jacusa2Helper is the JACUSA tidy data that represents
JACUSA2 result file "jacusa2.out" and can be created via \code{jacusa2 <- read_jacusa("jacusa2.out")}.
}
\section{input/output functions}{

The functions read_result, and write_result facilitate input and output operations on JACUSA (1.x and 2.x) 
output files.

See:
\itemize{
  \item read_result
  \item write_result
}
}

\section{get functions}{

This functions calculate TODO
}

\section{add functions}{

TODO
This functions calculate and add additional such as read depth or base call changes. 

See:
\itemize{
  \item add_bc_change
  \item add_bc_change_ratio
  \item add_coverage
}
}

\section{filter functions}{

This function set enables processing of JACUSA (1.x and 2.x) output files such as filtering by read coverage 
or enforcing a minimal number of variant base calls per sample.

See:
\itemize{
 \item filter_allele_count
  \item filter_call_stat
  \item filter_coverage
  \item filter_arrest_stat
  \item filter_robust_variants
  \item filter_min_variant_count
}
}

