% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{package}
\name{JACUSA2helper}
\alias{JACUSA2helper}
\alias{JACUSA2helper-package}
\title{JACUSA2helper: A package for post-processing JACUSA2 result files.}
\description{
JACUSA2helper: A package for post-processing JACUSA2 result files.
}
\section{Description}{

A package that provides the following categories of functions to post-process result files of JACUSA2:
\describe{
 \item{read/write}{Read and write JACUSA2 result files, e.g.: \code{read_result()}.}
 \item{add}{Adds some field to an existing JACUSA2 result object and return the modified object, e.g.: \code{add_sub()}.}
 \item{get}{TODO}
 \item{col(s)}{col(umns)TODO}
 \item{check}{Performs some checks on a result object, e.g.: \code{check_max_alleles()}.}
 \item{filter}{Will remove sites from a result object with some filtering criteria, e.g.: \code{filter_by_coverage()}}
 \item{plot}{Plots certain characteristics of a JACUSA2 result object.}
 \item{other}{Retrieves some field from an existing JACUSA2 result object. \code{base_type}}
}

The following methods from JACUSA2 are supported:
\describe{
 \item{call{1,2}}{calling variants from one condition or a comparison of conditions.}
 \item{pileup}{SAMtools like pileup.}
 \item{rt-arrest}{identification of transcription arrest events.}
 \item{lrt-arrest}{combination of variant discovery and read arrest events.}
}

One major new feature of JACUSA2 is the identification of read arrest events. 
In this method, the vector of base call is partitioned into read arrest and read through bases.

When working with stranded RNA-Seq data, inverting base calls is not necessary because
JACUSA2 will automatically invert Single End (SE) and Paired End (PE) depending on the
provided library type option "-P" UNSTRANDED|FR_FIRSTSTRAND|RF_SECOND_STRAND".

The central data structure in JACUSA2helper is the JACUSA2 result object that follows the 
tidy data approach to feature easy interaction with dplyr and ggplot2.
A JACUSA2 result object can be created via \code{result <- read_result("jacusa2.out")} and is 
currently represented as a tibble. Furthermore, JACUSA2helper supports the analysis of several related
JACUSA2 result files via \code{results <- read_results(files, meta_conditions)} where \code{meta_conditions} is a 
vector of character strings that provides a descriptive name for each file in \code{files}.

Check \code{vignette(TODO)} for a general introduction and \code{vignette(TODO meta conditions)} for details about meta conditions.
}

\section{site}{

A site is defined as a location using: contig, start, end, and strand information.
See:
\describe{
  \item{group_by_site}{Groups JACUSA2 result object by coordinates and optional columns.}
}
}

\section{read/write functions}{

See:
\describe{
  \item{read_result}{Reads and unpacks a JACUSA2 result file and creates a result object.}
  \item{read_results}{Allows to combine multiple result files and distinguish them with meta conditions.}
  \item{write_result}{This will pack result object and write its contents back to a file.}
  \item{write_bedGraph}{Writes a vector of values as bedGraph file.}
}
}

\section{add functions}{

See:
\describe{
  \item{add_arrest_rate}{Adds arrest rate to JACUSA2 result object.}
  \item{add_sub}{Adds base substitution column to JACUSA2 result object.}
  \item{add_sub_ratio}{Adds base substitution ratio for all bases to a JACUSA2 result object.}
  \item{add_TODOnon_ref_base2bc_ratio}{Adds non reference base ratio to JACUSA2 result object.}
  \item{add_summary}{Adds a summary of a column in a JACUSA2 result object.}
}
}

\section{check function}{

See:
\describe{
  \item{check_max_alleles}{Checks if site contains less or equal number of allowed alleles.}
}
}

\section{filter functions}{

This function set enables filtering by read coverage or 
enforcing a minimal number of variant base calls per sample.

See:
\describe{
 \item{filter_by_allele_count}{Filters sites by the number of alleles per site.}
  \item{filter_by_coverage}{Filters sites by read coverage (total, per replicates, etc.)}
  \item{filter_by_filter_info}{Filters sites by artefacts.}
  \item{filter_by_max_score}{Retains sites by maximal score.}
  \item{filter_by_min_score}{Retains sites by minimal score.}
  \item{filter_by_robust_arrest_events}{Retains sites that contain an arrest event in all replicates in at least one condition.}
  \item{filter_by_robust_variants}{Retains sites that contain a variant base in all replicates in at least one condition.}
}
}

