% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{package}
\name{JACUSA2helper}
\alias{JACUSA2helper}
\alias{JACUSA2helper-package}
\title{JACUSA2helper: A package for post-processing JACUSA2 result files.}
\description{
The JACUSA2helper package provides the following categories of important functions :
\itemize{
 \item[read/write] Read and write JACUSA2.
 \item[add] This will add some key to an existing JACUSA2 object and return the modified object.
 \item[check] This will perform some checks on JACUSA2 objects
 \item[filter] This will remove sites from a JACUSA2 object with some filtering criteria.
 \item[get] TODO
}
to post-process result files of JACUSA2.
}
\details{
The following methods are supported:
\itemize{
 \item[call{1,2}] calling variants from one condition or condition comparisons.
 \item[pileup] SAMtools like pileup
 \item[rt-arrest] identification of RNA modifications by reverse transcription arrest.
 \item[lrt-arrest] combination of varidant discovery and read arrest sites. 
}

One major new feature of JACUSA 2.x is the identificaion of RNA modifications by reverse
transcription arrest.
tidy data, dplyr

When working with stranded RNA-Seq data, inverting base calls is not necessary because
JACUSA2 will automatically invert Single End and Paired End depending on the provided 
library type option "-P" UNSTRANDED|FR_FIRSTSTRAND|RF_SECOND_STRAND".

The central data structure in JACUSA2helper is the JACUSA2 object that follows the 
tidy data approach to feature easy interplay with dplyr and ggplot2.
An JACUSA2 object can be created via \code{jacusa2 <- read_jacusa("jacusa2.out")} and is 
currently represented as a data.frame.
}
\section{read/write functions}{

The functions read_result, and write_result facilitate input and output operations on JACUSA2 
output files.

See:
\itemize{
  \item read_result
  \item write_result
}
}

\section{add functions}{

TODO

See:
\itemize{
  \item add_coverage
  \item add_ref_base2bc
  \item add_ref_base2bc_ratio
}
}

\section{check functions}{

TODO

See:
\itemize{
  \item chech_max_alleles
}
}

\section{filter functions}{

This function set enables filtering by read coverage or 
enforcing a minimal number of variant base calls per sample.

See:
\itemize{
 \item filter_by_allele_count
  \item filter_by_coverage
  \item filter_robust_variants
  \item filter_by_min_score
  \item filter_by_max_score
}

#' @section get functions:
TODO

See:
\itemize{
  \item get_bc_change
  \item get_bc_change_ratio
  \item get_variant_count
}
}

