% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ref-base2bc.R
\name{merge_ref_base2bc}
\alias{merge_ref_base2bc}
\title{Merges base change entries.}
\usage{
merge_ref_base2bc(ref_base2bc)
}
\arguments{
\item{ref_base2bc}{vector of base changes.}
}
\value{
vector of merged base changes.
}
\description{
TODO
}
