% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-ref-ratio.R
\name{non_ref_ratio}
\alias{non_ref_ratio}
\title{Calculate non reference ratio from base counts and reference base.}
\usage{
non_ref_ratio(ref, bases)
}
\arguments{
\item{ref}{string vector of reference bases.}

\item{bases}{nx4 matrix or data frame of base call counts.}
}
\value{
numeric vector of non reference substitution ratios for \code{base_type}.
}
\description{
Retrives non reference ratio based on \code{ref} and base calls specified by \code{base_type}.
}
\examples{
ref <- c("A", "A", "A")
bases <- tibble::tribble(
  ~A, ~C, ~G, ~T,
  10,  0,  5,  5,
   0,  1,  1,  1,
   1,  1,  1,  1
)
str(non_ref_ratio(ref, bases))
}
