% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-ref-ratio.R
\name{non_ref_ratio_col}
\alias{non_ref_ratio_col}
\title{Column name for non reference ratio for \code{base_type}.}
\usage{
non_ref_ratio_col(base_type)
}
\arguments{
\item{base_type}{string specifiying base column. Default: bases.}
}
\value{
string the represents the column name for non reference ratio for \code{base_type}.
}
\description{
Column name for non reference ratio for \code{base_type}. 
\code{base_type} will be used to create the column name:
\code{non_ref_ratio[_base_type]}. Using \code{base_type = "bases"} 
will result in the field \code{non_ref_ratio}.
}
