% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{peak}
\alias{peak}
\title{Formats JACUSA2 result object as in JACUSA2 result output.}
\usage{
peak(result)
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}
}
\value{
object structured as in JACUSA2 result format
}
\description{
Allows to peak at a result object and provides a user friendly format.
Converts data from long to wide format. @seealso \code{pack()}
}
