% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf_column}
\alias{plot_ecdf_column}
\title{Plot TODO}
\usage{
plot_ecdf_column(result, column)
}
\arguments{
\item{result}{object \code{read_results()}}

\item{column}{to be used for plotting}
}
\value{
Plot TODO
}
\description{
TODO
}
