% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ecdf-coverage.R
\name{plot_ecdf_coverage}
\alias{plot_ecdf_coverage}
\title{Plot cumulative read coverage.}
\usage{
plot_ecdf_coverage(result, data_desc, coverage_column = "coverage")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{data_desc}{vector of characters strings providing details about the sample.}

\item{coverage_column}{character sting specifing the name of the column.}
}
\value{
ggplot2 plot object.
}
\description{
Plots cumulative distribution of read coverate for single and multiple JACUSA2 result objects.
}
