% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ref-base2bc.R
\name{plot_ref_base2bc}
\alias{plot_ref_base2bc}
\title{Plot distribution of base changes.}
\usage{
plot_ref_base2bc(result, column = "ref_base2bc_column")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{column}{character string to be used for plotting.}
}
\value{
ggplot2 plot object.
}
\description{
TODO
}
\details{
The user is responsible for correct filtering and grouping of the 
\code{result} object.
\code{data_desc} is used to provide a descriptive name for each observation.

TODO make column replicate optional
}
