% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sub.R
\name{plot_sub}
\alias{plot_sub}
\title{Plot distribution of base substitutions.}
\usage{
plot_sub(result, column = "sub")
}
\arguments{
\item{result}{object created by \code{read_result()} or \code{read_results()}.}

\item{column}{character string to be used for plotting.}
}
\value{
ggplot2 plot object.
}
\description{
Plots the distribution of base substitutions.
The user is responsible for correct filtering and grouping of the 
\code{result} object.
}
