% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_result}
\alias{read_result}
\title{Read JACUSA2 result file}
\usage{
read_result(file, cond_desc = c(), unpack = FALSE, ...)
}
\arguments{
\item{file}{String that represents the file name of the JACUSA2 output.}

\item{cond_desc}{Vector of strings that represent names/descriptions for conditions.}

\item{unpack}{Boolean indicates if info column should be processed.}

\item{...}{parameters forwarded to `data.table::fread`}
}
\description{
\code{read_result()} reads data that was generated by JACUSA2 and creates a JACUSA2 result object.
}
