% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-result.R
\name{read_result}
\alias{read_result}
\title{Reads JACUSA2 result file and creates a result object.}
\usage{
read_result(file)
}
\arguments{
\item{file}{string that represents the filename of the JACUSA2 output.}
}
\description{
\code{read_result()} Reads data that was generated by JACUSA2 and creates a JACUSA2 result object.
}
