% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_result}
\alias{read_result}
\title{Read JACUSA2 result file}
\usage{
read_result(
  file,
  cond_desc = c(),
  unpack = FALSE,
  tmpdir = tempdir(),
  showProgress = FALSE,
  ...
)
}
\arguments{
\item{file}{String that represents the file name of the JACUSA2 output or an URL.}

\item{cond_desc}{Vector of strings that represent names/descriptions for conditions.}

\item{unpack}{Boolean or vector of fields to be unpacked.}

\item{tmpdir}{String @see `data.table::fread`.}

\item{showProgress}{Boolean @see `data.table::fread`.}

\item{...}{parameters forwarded to `data.table::fread`.}
}
\description{
\code{read_result()} reads data that was generated by JACUSA2 and creates a JACUSA2 result object.
}
