% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-result.R
\name{read_result}
\alias{read_result}
\title{Read JACUSA2 result file}
\usage{
read_result(file, showProgress = TRUE)
}
\arguments{
\item{file}{String that represents the filename of the JACUSA2 output.}

\item{showProgress}{Boolean indicates if progress should be monitored.}
}
\description{
\code{read_result()} Reads data that was generated by JACUSA2 and creates a JACUSA2 result object.
}
