% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_result}
\alias{read_result}
\title{Read JACUSA result file}
\usage{
read_result(file, showProgress = TRUE)
}
\arguments{
\item{file}{String represents the filename of the JACUSA2 output.}

\item{showProgress}{Boolean indicates if a progress should be monitored.}
}
\description{
\code{read_result()} Reads data that was generated by JACUSA2 and create a special data.frame.
}
