% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_results}
\alias{read_results}
\title{Read multiple related JACUSA2 results}
\usage{
read_results(files, meta_conds, cond_descs, unpack = FALSE, cores = 1)
}
\arguments{
\item{files}{Vector of files.}

\item{meta_conds}{Vector of string vectors that explain files.}

\item{cond_descs}{Vector of strings that represent names/descriptions for conditions.}

\item{unpack}{Boolean indicates if info column should be processed.}

\item{cores}{Integer defines how many cores to use.}
}
\description{
Read multiple related JACUSA2 results.
}
