% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-result.R
\name{read_results}
\alias{read_results}
\title{Read multiple related JACUSA2 results}
\usage{
read_results(files, meta_conditions)
}
\arguments{
\item{files}{vector of files}

\item{meta_conditions}{vector of string describing each}
}
\value{
combined JACUSA2 result object
}
\description{
TODO
}
