% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-by-robust-arrest-event.R
\name{robust_arrest_events}
\alias{robust_arrest_events}
\title{TODO}
\usage{
robust_arrest_events(condition, arrest_cov, through_cov)
}
\arguments{
\item{condition}{TODO}

\item{arrest_cov}{TODO}

\item{through_cov}{TODO}
}
\value{
TODO
}
\description{
This is intended to be used within a \code{group_by_site()} statement. TODO
}
\examples{
TODO
}
