% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rrd}
\alias{rrd}
\title{Subset of RRDs detected by JACUSA2 in HEK-293 ADAR KD and untreated cells}
\format{a tibble with 19 elements:
\itemize{
	\item contig: Character string representing the contig of the variant
	\item start: Numeric position of variant (>=0)
	\item end: Numeric corresponds to "start + 1"
	\item name: Character string. Currently, name of used method (\emph{call-\{1,2\}}, \emph{pileup}, \emph{rt-arrest}, or \emph{lrt-arrest})
	\item score: Numeric value representing the test-statistc. Higher values indicate more divergent pileups
	\item strand: Character representing strand information; "+", "-", or "."(no strand information available)
	\item info: Character string separated with ";" provding additional data for this specific site. Empty field is equal to "*"
	\item filter: ";"-separated character string showing feature filter information. Empty field is equal to "*"
	\item ref: Character "A", "C", "G", "T", or "N" representing the reference base for this site - inverted when strand is "-".
	\item bases: Numeric tibble representing counts for A, C, G, and T base calls.
	\item cond: Numeric value: "1"(here: ADAR KD) or "2"(here: unreated)
	\item repl Numeric value: "1" or "2" for all conditions
	\item bc: Observed base calls for this site. Combination of: "A", "C", "G", and "T"
   \item cov: Numeric value indicating the read coverage for this site
}}
\usage{
data(rrd)
}
\description{
Data extrated from:
Piechotta, M.; Wyler, E.; Ohler, U.; Landthaler, M. & Dieterich, C.
JACUSA: site-specific identification of RNA editing events from replicate sequencing data 
BMC Bioinformatics, Springer Nature, 2017 , 18
}
\references{
Piechotta, M.; Wyler, E.; Ohler, U.; Landthaler, M. & Dieterich, C.
JACUSA: site-specific identification of RNA editing events from replicate sequencing data 
BMC Bioinformatics, Springer Nature, 2017 , 18

A dataset containing a subset of RNA RNA differences (RRDs) identified by JACUSA2 in ADAR KD and untreated HEK-293 cells. 
The fields are as follows:
}
\keyword{datasets}
