% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rrd}
\alias{rrd}
\title{Subset of RRDs detected by JACUSA2 in HEK-293 ADAR KD and untreated cells}
\format{
a tibble with 18 elements:
\itemize{
   \item id: Character string representing a unique identifier - created from contig, start, [end], and strand.
	\item contig: Character string representing the contig of the variant
	\item start: Numeric position of variant (>=0)
	\item end: Numeric corresponds to "start + 1"
	\item name: Character string. Name of used method \emph{call-2}
	\item score: Numeric value representing the test-statistc. Higher values indicate more divergent pileups
	\item strand: Character representing strand information; "+", "-", or "."(no strand information available)
	\item bases: Numeric tibble with representing counts for A, C, G, and T base calls.
   \item cov: Numeric value indicating the read coverage for this site
   \item backtrack1: Character - indicator if backtracking was used for condition 1.
   \item backtrack2: Character - indicator if backtracking was used for condition 2.
   \item backtrackP: Character - indicator if backtracking was used for condition pooled condition.
   \item reset1: Character - indicator if default estimation was unstable with for condition 1.
   \item reset2: Character - indicator if default estimation was unstable with for condition 2.
   \item resetP: Character - indicator if default estimation was unstable with for pooled condition.
	\item info: Character string separated with ";" provding additional data for this specific site. Empty field is equal to "*"
	\item filter: ";"-separated character string showing feature filter information. Empty field is equal to "*"
	\item ref: Character "A", "C", "G", "T", or "N" representing the reference base for this site - inverted when strand is "-".
}
}
\usage{
data(rrd)
}
\description{
Data extrated from:
Piechotta, M.; Wyler, E.; Ohler, U.; Landthaler, M. & Dieterich, C.
JACUSA: site-specific identification of RNA editing events from replicate sequencing data 
BMC Bioinformatics, Springer Nature, 2017 , 18
}
\references{
Piechotta, M.; Wyler, E.; Ohler, U.; Landthaler, M. & Dieterich, C.
JACUSA: site-specific identification of RNA editing events from replicate sequencing data 
BMC Bioinformatics, Springer Nature, 2017 , 18

A dataset containing a subset of RNA RNA differences (RRDs) identified by JACUSA2 in ADAR KD and untreated HEK-293 cells. 
The fields are as follows:
}
\keyword{datasets}
