% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot}
\usage{
save_plot(filename, plot, ...)
}
\arguments{
\item{filename}{character vector}

\item{plot}{object to be saved}

\item{...}{parameters passed to ggplot2::ggsave}
}
\description{
Saves plot using ggplot2::ggsave.
}
