% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-plot.R
\name{save_plot}
\alias{save_plot}
\title{Saves plot.}
\usage{
save_plot(filename, plot, ...)
}
\arguments{
\item{filename}{character string naming a file for writing.}

\item{plot}{the ggplot2 plot object to be written.}

\item{...}{parameters passed to ggplot2::ggsave}
}
\description{
Wrapper to save plot using ggplot2::ggsave.
}
