% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-ratio.R
\name{sub_ratio}
\alias{sub_ratio}
\title{Calculates base substitution ratio (e.g.: editing frequency).}
\usage{
sub_ratio(ref, bases, bc = NULL)
}
\arguments{
\item{ref}{vector of reference bases.}

\item{bases}{matrix of observed base call counts.}

\item{bc}{vector strings that represents observed base calls. Default: "bases".}
}
\value{
vector of base callsubstitution ratios.
}
\description{
Calculates base substitution ratio (e.g.: editing frequency) for \code{ref} and base counts stored
in nx4 matrix \code{bases}.
\code{ref} must have only one reference base. A->G is okay, but AG->G is NOT allowed!
If \code{bc} is not provided, only one non-reference base in \code{bases} is allowed!
}
\examples{
ref <- c("A", "A", "T")
# Only one non-ref. base
bases <- matrix(c( 1,  0, 9, 0,
                   5,  0, 5, 0,
                   0, 10, 0, 0),
                ncol = 4, byrow = TRUE)
# show base substiution ratios for observed base calls
sub_ratio(ref, bases)

# > one non-ref. base
bases <- matrix(c( 1,  2, 9, 1,
                   5,  3, 5, 1,
                   2, 10, 2, 1),
                ncol = 4, byrow = TRUE)
# we are only interesected in A->G, T->C
bc <- c("G", "G", "C")
# show base substiution ratios for A->G, T->C
sub_ratio(ref, bases, bc)
}
