% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest-rate.R
\name{through_col}
\alias{through_col}
\title{Column name for through base counts for \code{suffix}.}
\usage{
through_col(suffix = NULL)
}
\arguments{
\item{suffix}{TODO}
}
\value{
string the represents the column name for through for \code{suffix}.
}
\description{
Column name for through base counts for \code{suffix}. 
\code{suffix} will be used a suffix to create the column name:
"through[_\code{suffix}]".
}
