% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{unpack_info}
\alias{unpack_info}
\title{Unpack info field}
\usage{
unpack_info(result, cond_count, keys = get_info_keys(result), cores = 1)
}
\arguments{
\item{result}{JACUSA2 result object.}

\item{cond_count}{integer Number of conditions.}

\item{cores}{integer Number of compute cores to use. Default: 1.}
}
\value{
returns a JACUSA2 result object with "info" unpacked.
}
\description{
Unpacks info field.
}
