% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{unpack_info}
\alias{unpack_info}
\title{Unpack info field}
\usage{
unpack_info(info, cond_count, keys = get_info_keys(info))
}
\arguments{
\item{info}{Vector of stings.}

\item{cond_count}{integer Number of conditions.}

\item{keys}{vector of string of keys to extract from "info".}
}
\value{
returns a JACUSA2 result object with "info" unpacked.
}
\description{
Unpacks info field.
}
