% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-bedGraph.R
\name{write_bedGraph}
\alias{write_bedGraph}
\title{Writes vector of values as bedGraph file.}
\usage{
write_bedGraph(file, contig, start, end, value)
}
\arguments{
\item{file}{String represents the filename of the BED graph file.}

\item{contig}{vector of character string represting the contig(s)}

\item{start}{vector of numericals zero indexed (>= 0)}

\item{end}{vector of numericals half opened: [start, end)}

\item{value}{vector of character strings or numericals representing value(s)}
}
\description{
Writes a bedGraph conform file with coordinates
defined by vectors: "contig", "start", and "end". According to:
https://genome.ucsc.edu/goldenPath/help/bedgraph.html
"start" and "end" are zero indexed and half opened. The caller needs to make sure 
that this is fullfilled.
}
