% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-write.R
\name{write_result}
\alias{write_result}
\title{Write a JACUSA list object to a file}
\usage{
write_result(jacusa2, file, extra = NULL)
}
\arguments{
\item{jacusa2}{List created by \code{read_jacusa()}.}

\item{file}{String is the filename to store the list.}

\item{extra}{Vector of strings that define additional elements from the list that 
will be stored in the file.}
}
\description{
\code{write_result} Stores a list of sites in a file.
}
